% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Breast.fit.R
\docType{data}
\name{Breast.fit}
\alias{Breast.fit}
\title{Fit from the Yates et al. cohort of Breast cancers, with REVOLVER. This file contains also clustering assignments and jacckknife results.}
\format{An S3 object of class "rev_cohort_fit".}
\usage{
data(Breast.fit)
}
\description{
Driver mutations annotated in the original paper.
}
\examples{
data(Breast.fit)
print(Breast.fit)

}
\references{
Yates et al. (2015) Nat Med. 2015 Jul;21(7):751-9.
}
\keyword{datasets}
