% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{CCF_clusters}
\alias{CCF_clusters}
\title{Extract cluster-level CCF/binary data for a patient}
\usage{
CCF_clusters(x, p)
}
\arguments{
\item{x}{a REVOLVER object of class "rev_cohort" or "rev_cohort_fit"}

\item{p}{a patient}
}
\value{
CCF/binary data for \code{p}
}
\description{
Extract cluster-level CCF/binary data for a patient
}
\examples{
data(TRACERx_cohort)
CCF_clusters(TRACERx_cohort, 'CRUK0001')
}
