% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Cluster}
\alias{Cluster}
\title{Get clustering information for patients.}
\usage{
Cluster(x, patients = x$patients)
}
\arguments{
\item{x}{A REVOLVEr cohort with fits and clusters.}

\item{patients}{Patients to use, all by default.}
}
\value{
A tibble with the required clustering assignments
}
\description{
For a set of `patients` - all by default - 
extract the clustering information computed by REVOVLER.
}
\examples{
TODO
}
