% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_clone_trees.R
\name{compute_clone_trees}
\alias{compute_clone_trees}
\title{Compute CCF-based phylogenies for a \code{REVOLVER} cohort.}
\usage{
compute_clone_trees(x, patients = Stats_cohort(x) \%>\% pull(patientID),
  overwrite = FALSE, ...)
}
\arguments{
\item{x}{A \code{REVOLVER} cohort.}

\item{patients}{A set of patient ids in the cohort, for which the
phylogenies are created.}
}
\value{
A modififed object of class \code{"rev_cohort"} with available
phylogeneis for \code{patients}.
}
\description{
Create clone tree phylogenies from the package
\code{ctree}, fitting the input data given to \code{REVOLVER}. 
A set of patient ids can be given as input, by default all are
used. A parmeter controls if you want to overwrite the trees
of the patient, where they already computed. Please refer to the
parameters of the \code{ctrees} constructor rom package \code{ctree} 
for the input parameters that you can use to tune the construction.
}
\examples{
data(TRACERx_cohort)

print(TRACERx_cohort)

# We use the standard parameters with overwrite = TRUE
# otherwise the tool returns retuns the trees
# already available
TRACERx_cohort = revolver_compute_phylogenies(
   TRACERx_cohort, 
   "CRUK0002",
   overwrite = TRUE
    )
    
Phylo(TRACERx_cohort, 'CRUK0002', rank = 1)
}
