% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mutation_trees.R
\name{compute_mutation_trees}
\alias{compute_mutation_trees}
\title{Compute binary mutation trees for a patient (phylogenies).}
\usage{
compute_mutation_trees(
  x,
  patients = Stats_cohort(x) \%>\% pull(patientID),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{REVOLVER} cohort.}

\item{patients}{A set of patient ids in the cohort, for which the
phylogenies are created.}

\item{overwrite}{If \code{TRUE}, recompute trees even if already available.}

\item{...}{Additional parameters forwarded to the \code{mtree} constructor.}
}
\value{
A \code{REVOLVER} cohort with mutation tree phylogenies for the
requested \code{patients}.
}
\description{
Create mutation tree phylogenies from the package
\code{mtree}, fitting the input data given to \code{REVOLVER}.
A set of patient ids can be given as input, by default all are
used. A parameter controls if you want to overwrite the trees
of the patient, where they already computed. Please refer to the
parameters of the \code{mtree} constructor from package \code{mtree}
for the input parameters that you can use to tune the construction.
}
\examples{
\dontrun{
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')
x = compute_mutation_trees(TRACERx_NEJM_2017_REVOLVER)
}
}
\seealso{
Other Cohort creation: 
\code{\link{CCF_parser}()},
\code{\link{compute_clone_trees}()},
\code{\link{input_custom_trees}()},
\code{\link{revolver_check_cohort}()},
\code{\link{revolver_cohort}()}
}
\concept{Cohort creation}
