% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRC.cohort.R
\docType{data}
\name{CRC.cohort}
\alias{CRC.cohort}
\title{Colorectal Cancer Data by Cross et al., in the cohort format by REVOLVER.}
\format{An S3 object of class "rev_cohort".}
\usage{
data(CRC.cohort)
}
\description{
Driver mutations detected by Cross et al. in a cohort of colorecta adenomas and carcinomas.
This dataset has already been wrapped to a REVOLVER cohort object
}
\examples{
data(CRC.cohort)
print(CRC.cohort)

}
\references{
Cross et al. (2018) to appear in Nature Ecology and Evolution.
}
\keyword{datasets}
