% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_getters.R
\name{D}
\alias{D}
\title{Extract CCF/binary data for a patient}
\usage{
D(x, p)
}
\arguments{
\item{x}{a REVOLVER object of class "rev_cohort" or "rev_cohort_fit"}

\item{p}{a patient}
}
\value{
CCF/binary data for \code{p}
}
\description{
Extract CCF/binary data for a patient
}
\examples{
data(CRC.cohort)
CCF(CRC.cohort, 'adenoma_3')
}
