% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DET_index.R
\name{DET_index}
\alias{DET_index}
\title{Title}
\usage{
DET_index(x, drivers = x$variantIDs.driver, min.occurrences = 0)
}
\arguments{
\item{x}{}

\item{drivers}{}

\item{min.occurrences}{Extract the information transfer available for a patient.}

\item{x}{A \code{REVOLVER} cohort with fits.}

\item{drivers}{The list of drivers to compute the DET for.}

\item{min.occurrences}{The minimum number of occurrences for
a trajectory to be considered, zero by default. See also 
function \code{\link{plot_penalty}}.}
}
\value{
The information transfer for the tree of a patient
}
\description{
This function computes the index of Divergent Evolutionary 
Trajectories for the fit of the input cohort.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')
 
# Get the transfer among drivers, top-ranking tree
DET_index(TRACERx_NEJM_2017_REVOLVER)
}
\seealso{
Other Getters: \code{\link{CCF_clusters}},
  \code{\link{CCF}}, \code{\link{Clonal_cluster}},
  \code{\link{Data}}, \code{\link{Drivers}},
  \code{\link{ITransfer}}, \code{\link{Phylo}},
  \code{\link{Samples}}, \code{\link{Subclonal}},
  \code{\link{Truncal}}
}
\concept{Getters}
