% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Drivers}
\alias{Drivers}
\title{Extract CCF/binary driver data for a patient}
\usage{
Drivers(x, p)
}
\arguments{
\item{x}{a REVOLVER object of class "rev_cohort" or "rev_cohort_fit"}

\item{p}{a patient}
}
\value{
CCF/binary drivers data for \code{p}
}
\description{
Extract CCF/binary driver data for a patient
}
\examples{
data(CRC.cohort)
Drivers(CRC.cohort, 'adenoma_3')
}
