% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_getters.R
\name{Fit}
\alias{Fit}
\title{Extract fitted model for a patient}
\usage{
Fit(x, p)
}
\arguments{
\item{x}{a REVOLVER object of class "rev_cohort_fit"}

\item{p}{a patient}
}
\value{
fitted model for  \code{p}
}
\description{
Extract fitted model for a patient
}
\examples{
data(CRC.cohort)

Fit(revolver_fit(CRC.cohort), 'adenoma_3')
}
