% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{gradient_palette}
\alias{gradient_palette}
\title{Gradient palette.}
\usage{
gradient_palette(n = NULL)
}
\arguments{
\item{n}{The number of required colours. By default this is NULL.}
}
\value{
A function that applied to a number returns the required colours.
}
\description{
ColorRamp palette extrapolating \code{YlGnBu} palette of
\code{RColorBrewer} package. The original \code{YlGnBu} palette is
sampled for 9 colours.
}
\examples{
gradient_palette(2)
}
