% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_custom_trees.R
\name{input_custom_trees}
\alias{input_custom_trees}
\title{Add a custom set of precomputed trees for a patient.}
\usage{
input_custom_trees(x, patient, trees, scores)
}
\arguments{
\item{x}{An object of class \code{"rev_cohort"}}

\item{patient}{A patient in the cohort, for which phylogenies are created}

\item{trees}{A list a of precomputed adjacency matrix to describe the custom trees.}

\item{scores}{A list a of precomputed scores to describe the custom trees.}
}
\value{
A modififed REVOLVER cohort with available phylogeneis for \code{patient}.
}
\description{
TODO
}
\examples{
TODO
}
