% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{ITransfer}
\alias{ITransfer}
\title{Extract the information transfer for a patient's tree}
\usage{
ITransfer(x, p, rank = 1, type = "drivers", data = "trees")
}
\arguments{
\item{x}{A REVOLVER cohort}

\item{p}{A patient}

\item{rank}{The rank of the tree to extract}

\item{type}{Either `clones` or `drivers`.}

\item{data}{Either `trees` or `fit`.}
}
\value{
The information transfer for a patient's tree
}
\description{
Extract the information transfer for a patient's tree
}
\examples{
data(CRC.cohort)
Phylo(CRC.cohort, 'adenoma_3')
}
