% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Jackknife_cluster_stability}
\alias{Jackknife_cluster_stability}
\title{Get jackknifed cluster stability.}
\usage{
Jackknife_cluster_stability(x)
}
\arguments{
\item{x}{A REVOLVER cohort with fits, clusters and jackknife results.}
}
\value{
A named vector.
}
\description{
After jackknife computations via \code{\link{revolver_jackknife}},
this function extracts a cluster-specific stability obtained as the median
co-clustering probability of the patients in the cluster.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

Jackknife_cluster_stability(TRACERx_NEJM_2017_REVOLVER)
}
\seealso{
Other Jackknife statistics: \code{\link{Jackknife_patient_coclustering}},
  \code{\link{Jackknife_trajectories_stability}}
}
\concept{Jackknife statistics}
