% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Jackknife_trajectories_stability}
\alias{Jackknife_trajectories_stability}
\title{Get jackknifed trajectories stability.}
\usage{
Jackknife_trajectories_stability(x)
}
\arguments{
\item{x}{A REVOLVER cohort with fits, clusters and jackknife results.}
}
\value{
A tibble.
}
\description{
After jackknife computations via \code{\link{revolver_jackknife}},
this function extracts the trajectories stability obtained as the number of times
(normalized and non) a trajectory is inferred across resamples.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

Jackknife_trajectories_stability(TRACERx_NEJM_2017_REVOLVER)
}
\seealso{
Other Jackknife statistics: \code{\link{Jackknife_cluster_stability}},
  \code{\link{Jackknife_patient_coclustering}}
}
\concept{Jackknife statistics}
