% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Phylo}
\alias{Phylo}
\title{Extract phylogenetic or mutation trees for a patient}
\usage{
Phylo(x, p, rank = NULL, data = "trees")
}
\arguments{
\item{x}{a REVOLVER object of class "rev_cohort" or "rev_cohort_fit"}

\item{p}{a patient}

\item{data}{Either `trees` or `fit`.}
}
\value{
phylogenetic or mutation trees data for \code{p}
}
\description{
Extract phylogenetic or mutation trees for a patient
}
\examples{
data(TRACERx_cohort)

# Get all the trees for a patient
Phylo(TRACERx_cohort, 'CRUK0002')

# Get a specific tree for a patient
Phylo(TRACERx_cohort, 'CRUK0002', rank = 2)
# Get the fit tree for a patient
}
