% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R, R/revolver_cohort.R
\name{plot.rev_cohort}
\alias{plot.rev_cohort}
\title{Plot data and trees for a REVOLVER cohort.}
\usage{
\method{plot}{rev_cohort}(x, ...)

\method{plot}{rev_cohort}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"rev_cohort"}.}

\item{patients}{The patients to plot, default is all the one available.}

\item{max.phylogenies}{How many trees should be computed for each patient.}

\item{cex}{Scale cex for graphics.}

\item{x}{A REVOLVER cohort.}

\item{cex}{Cex of the plot.}
}
\value{
nothing

A \code{ggplot2} object of the plot.
}
\description{
Plot data and trees for a REVOLVER cohort.

Cohort summary scatterplot.
}
\details{
Iterative plotting functions that scans a cohort and runs \code{\link{revolver_report_patient}}
on a set of patients (default all).

Returns a scatterplot of the tumour mutational burden. at the
clonal and subclonal level. Each dot is sized by the number of
clones with drivers, and coloured by the numnber of drivers.
}
\examples{
data(Breast.fit)
plot.rev_cohort(Breast.fit, patients = Breast.fit$patients[1:10])
data(Breast.fit)
}
