% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_fit.R
\name{plot.rev_cohort_fit}
\alias{plot.rev_cohort_fit}
\title{Plot REVOLVER fits for a subset of patients in the cohort.}
\usage{
\method{plot}{rev_cohort_fit}(x, patients = names(x$fit$phylogenies),
  plot.stat = TRUE, palette = "Set1", out.file = "REVOLVER-cohort.pdf",
  layout = paste(length(patients), "x1", sep = ""), alpha = 0.7,
  verbose = FALSE, ...)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object}

\item{patients}{The patients to plot, default is all the one available.}

\item{plot.stat}{Set it to \code{TRUE} to visualize also some statistics about each tree}

\item{palette}{RColorBrewer palette for colours of each model.}

\item{out.file}{Output PDF file, default \code{"REVOLVER-cohort.pdf"}.}

\item{layout}{The layout of to arrange the plots.}

\item{alpha}{Transparency}

\item{verbose}{Verbose output.}

\item{...}{Extra parameters}
}
\value{
None
}
\details{
This function performs the plots for each one of a set of input patient IDs. Each plot
contain the tree fit, its expanded version and the information transfer.
}
\examples{
#' data(CRC.cohort)
fit = revolver_fit(CRC.cohort)
plot(fit, "adenoma_3")
}
