% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_phylogeny.R
\name{plot.rev_phylo}
\alias{plot.rev_phylo}
\title{S3 method calling \code{\link{revolver_plt_tree}}.}
\usage{
\method{plot}{rev_phylo}(x, file = NA, palette = "Set1", cex = 1,
  alpha = 0.7)
}
\arguments{
\item{x}{An object of class \code{"rev_phylo"}.}

\item{file}{Output file, or \code{NA}.}

\item{palette}{RColorBrewer palette to colour clusters.}

\item{cex}{Cex for the graph.}

\item{alpha}{Transparency.}

\item{verbose}{Output.}

\item{...}{Extra parameters}
}
\value{
Nothing
}
\description{
S3 method calling \code{\link{revolver_plt_tree}}.
}
\examples{
data(CRC.cohort)
plot(CRC.cohort$phylogenies[['adenoma_3']][[1]])
}
