% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Title}
\usage{
plot_clusters(x, cluster_palette = revolver:::distinct_palette_few,
  driver_palette = revolver:::distinct_palette_many,
  CCF_palette = revolver:::gradient_palette, cutoff_drivers = 10,
  cutoff_trajectories = 10, arrow.symbol = "-->", ...)
}
\arguments{
\item{x}{}

\item{cluster_palette}{}

\item{driver_palette}{}

\item{CCF_palette}{}

\item{cutoff_drivers}{}

\item{cutoff_trajectories}{}

\item{arrow.symbol}{}

\item{...}{}
}
\description{
Title
}
