% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot the data for a patient.}
\usage{
plot_data(x, patient, ...)
}
\arguments{
\item{x}{A REVOLVER cohort object.}

\item{patient}{A patient id.}

\item{...}{Extra parameters passed to all the used plotting functions.}
}
\value{
A figure assembled with \code{ggpubr}.
}
\description{
This function creates a complex plot for the
data of a patient, assembling plots returned from thw following
functions: \code{\link{plot_data_clusters}}, 
\code{\link{plot_data_clone_size}}, \code{\link{plot_data_mutation_burden}}
and \code{plot_patient_oncoprint}.
}
\examples{
TODO
}
