% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_clone_size.R
\name{plot_data_clone_size}
\alias{plot_data_clone_size}
\title{Plot the clone size data for a patient.}
\usage{
plot_data_clone_size(x, patient, alpha_level = 0.05, cex = 1, ...)
}
\arguments{
\item{x}{A REVOLVER cohort object.}

\item{patient}{A patient id.}

\item{alpha_level}{Alpha level for the test, default is 0.05.}

\item{cex}{Cex of the plot.}

\item{...}{Extra parameters, not used.}
}
\value{
A \code{ggplot} plot.
}
\description{
This function creates a \code{ggplot}
barplot of the clone size values foe each clone in a patient's data.
The size of a clone is defined as the number of mutations assigned 
to it. The barplot is annotated to report wether a subclone with a driver
is significantly larger than the expected size for a subclone without
driver. To carry out this test subclones without drivers are used to
estimate the parameters of a univariate Gaussian distribution (mean
and standard deviation), the p-value is then computed from the fit
distribution through the `pnorm` function. The confidence level for
the test can be passed as parameter.
}
\examples{
TODO
}
