% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_clusters.R
\name{plot_data_clusters}
\alias{plot_data_clusters}
\title{Plot the clusters data for a patient.}
\usage{
plot_data_clusters(x, patient, cex = 1, ...)
}
\arguments{
\item{x}{A REVOLVER cohort object.}

\item{patient}{A patient id.}

\item{cex}{Cex of the plot.}

\item{...}{Extra parameters, not used.}
}
\value{
A \code{ggplot} plot.
}
\description{
This function creates a \code{ggplot}-style
heatmap of the mean CCF of each clone in a patient's data.
The heatmap is annotated for the drivers status of each
clone (with/ without driver). The CCF values are used to
colour the heatmap (`NA` values are in light gray).
}
\examples{
TODO
}
