% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_histogram.R
\name{plot_data_histogram}
\alias{plot_data_histogram}
\title{Plot the data histogram for a patient.}
\usage{
plot_data_histogram(x, patient, cex = 1, ...)
}
\arguments{
\item{x}{A REVOLVER cohort object.}

\item{patient}{The id of a patient.}

\item{cex}{Cex of the plot.}

\item{...}{Extra parameters, not used.}
}
\value{
A \code{ggplot} plot.
}
\description{
Plot the CCF histogram of each one of the samples available for
a patient, facetting the sample variable.
}
\examples{
TODO
}
