% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_mutation_burden.R
\name{plot_data_mutation_burden}
\alias{plot_data_mutation_burden}
\title{Plot the mutation burden for a patient.}
\usage{
plot_data_mutation_burden(x, patient, cex = 1, ...)
}
\arguments{
\item{x}{A REVOLVER cohort object.}

\item{patient}{A patient id.}

\item{cex}{Cex of the plot.}

\item{...}{Extra parameters, not used.}
}
\value{
A \code{ggplot} plot.
}
\description{
This function creates a scatterplot
of a full cohort annotating clonal versus subclonal
mutational burden, colouring each point by the 
number of total drivers in the patient, and using
as shape the number of clones with drivers.
}
\examples{
TODO
}
