% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_oncoprint.R
\name{plot_data_oncoprint}
\alias{plot_data_oncoprint}
\title{Plot the oncoprint for a patient.}
\usage{
plot_data_oncoprint(x, patient,
  clusters_palette = revolver:::distinct_palette_many, ...)
}
\arguments{
\item{x}{A REVOLVER cohort object.}

\item{patient}{The id of a patient.}

\item{clusters_palette}{A palette function that should return the colour of
an arbitrary number of clusters.}

\item{...}{Extra parameters, not used.}
}
\value{
A figure assembled with \code{ggpubr} which combines multiple
\code{ggplot} tile plots.
}
\description{
Plot the oncoprint for a patient.
}
\examples{
TODO
}
