% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendrogram.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot the dendrogram of REVOLVER"s clusters.}
\usage{
plot_dendrogram(x, cluster_palette = distinct_palette_few)
}
\arguments{
\item{x}{A \code{REVOLVER} object with fits and clusters.}

\item{clusters_palette}{A palette function that should return the colour of
an arbitrary number of clusters.}
}
\value{
A \code{ggplot} plot.
}
\description{
Plot the dendrogram of REVOLVER"s clusters, where leaves are patient
ids, coloured by cluster.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

plot_dendrogram(TRACERx_NEJM_2017_REVOLVER)
}
\seealso{
Other Plotting functions: 
\code{\link{distinct_palette_few}()},
\code{\link{distinct_palette_many}()},
\code{\link{gradient_palette}()},
\code{\link{plot_DET_index}()},
\code{\link{plot_clusters}()},
\code{\link{plot_drivers_clonality}()},
\code{\link{plot_drivers_graph}()},
\code{\link{plot_drivers_occurrence}()},
\code{\link{plot_jackknife_cluster_stability}()},
\code{\link{plot_jackknife_coclustering}()},
\code{\link{plot_jackknife_trajectories_stability}()},
\code{\link{plot_patient_CCF_histogram}()},
\code{\link{plot_patient_data}()},
\code{\link{plot_patient_mutation_burden}()},
\code{\link{plot_patient_oncoprint}()},
\code{\link{plot_patient_trees_scores}()}
}
\concept{Plotting functions}
