% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DET_index.R
\name{plot_DET_index}
\alias{plot_DET_index}
\title{Plot the DET index.}
\usage{
plot_DET_index(x, cex = 1, ...)
}
\arguments{
\item{x}{A REVOLVER cohort with fits.}

\item{cex}{Cex of the plot.}

\item{...}{Parmeters passed to function \code{DET_index} in order to compute
the index value.}
}
\value{
A \code{ggplot} object for the plot
}
\description{
Plot the index of Divergent Evolutionary Trajectories, for a set of drivers
using function \code{DET_index}. The plot is a barplot with colours reflecting
the number of distinct incoming edges in each driver (species), and the height
reflecting the actual DET index value.
}
\examples{
data(Breast.fit)
plot_DET_index(Breast.fit, min.occurrences = 5)
}
