% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph_drivers.R
\name{plot_graph_drivers}
\alias{plot_graph_drivers}
\title{Plot graph-alike summary statistics for the cohort drivers.}
\usage{
plot_graph_drivers(x, drivers = x$variantIDs.driver,
  min.occurrences = 0, alpha_level = 0.05, ...)
}
\arguments{
\item{x}{A REVOLVER object with fits.}

\item{drivers}{The list of drivers to consider, all by default.
See also function \code{\link{plot_penalty}}.}

\item{min.occurrences}{The minimum number of occurrences for
a trajectory to be considered, zero by default. See also 
function \code{\link{plot_penalty}}.}

\item{alpha_level}{The significance level for the enrichment Fisher test.}

\item{...}{Extra parameters passed to the \code{create_layout} function
by \code{ggraph}. For instance, passing \code{algorithm = 'kk'} and 
\code{layout = 'igraph'} the `igraph` layout `kk` will be adopted.}
}
\value{
A `ggplot` object of the plot.
}
\description{
Plot a graph with driver genes and annotate with 
different summary statistics for the trajectories that 
involve the drivers. This visualisation shows the frequency
of the driver in the cohort (node size), the penalty for
each pair of odering (edge thickness), the significance for
the pair of orderings as of a Fisher test (edge coloring)
and the overall heterogeneity upstream a driver as of the
DET index (node coloring). This function has parameters
to subset the computation to a list of predefined drivers,
or drivers associated to trajectories with a minimum
recurrence in the fits.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')
 
# Base plot, can be quite crowded
plot_graph_drivers(TRACERx_NEJM_2017_REVOLVER)

# Reduce the number of nodes cutting off low-frequencies one
plot_graph_drivers(TRACERx_NEJM_2017_REVOLVER, min.occurrences = 5)

# As above, but with a more stringent test
plot_graph_drivers(TRACERx_NEJM_2017_REVOLVER, min.occurrences = 5, alpha_level = 0.01)
}
\seealso{
Other Plotting functions: \code{\link{plot_CCF_histogram}},
  \code{\link{plot_DET_index}},
  \code{\link{plot_drivers_clonality}},
  \code{\link{plot_drivers_occurrence}},
  \code{\link{plot_patient_data}},
  \code{\link{plot_patient_mutation_burden}},
  \code{\link{plot_patient_oncoprint}},
  \code{\link{plot_trees_scores}}
}
\concept{Plotting functions}
