% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_information_transfer.R
\name{plot_information_transfer}
\alias{plot_information_transfer}
\title{Plot the information transfer of a tree.}
\usage{
plot_information_transfer(x, cex = 1,
  node_palette = colorRampPalette(RColorBrewer::brewer.pal(n = 9,
  "Set1")), tree_layout = "tree", ...)
}
\arguments{
\item{x}{A REVOLVER tree.}

\item{cex}{Cex of the plot.}

\item{node_palette}{A function that can return, for an input number,
a number of colours.}

\item{tree_layout}{Layout of this model, as of \code{ggraph}.}

\item{...}{Other parameters, not used in this case.}
}
\value{
A \code{ggplot} object for the plot
}
\description{
Compute via \code{ggraph} a graph for the
information transfer of a REVOLVER's model. The tree is coloured
so that drivers in the same clone have the same colour.
}
\examples{
data(TRACERx_cohort)

# Extract transfer
tree = Phylo(TRACERx_cohort, 'CRUK0002', rank = 1)

plot_information_transfer(tree)

# Change layout -- use igraph's "kk" layout
plot_information_transfer(tree, tree_layout = 'kk')
}
