% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jackknife_trajectories.R
\name{plot_jackknife_trajectories_stability}
\alias{plot_jackknife_trajectories_stability}
\title{Plot the patients' jackknife trajectory frequencies.}
\usage{
plot_jackknife_trajectories_stability(x, annotate_probability = 0.9,
  annotate_percentage = 0.2)
}
\arguments{
\item{x}{A \code{REVOLVER} cohort with fits, clusters and jackknife results available.}

\item{cluster_palette}{A palette of colours; must be a function that when applied to
a number it returns that number of colours.}
}
\value{
A \code{ggplot} figure.
}
\description{
This function plots a raster of the patients' jackknife co-clustering probability.
 
 The input matrix to this function can be obtained via \code{\link{Jackknife_patient_coclustering}};
 in the visualisation the patient ids are coloured and ordered according to the clusters in the
 input dataset, which can be obtained via \code{\link{Cluster}}.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

plot_jackknife_coclustering(TRACERx_NEJM_2017_REVOLVER)
}
\seealso{
Other Plotting functions: \code{\link{distinct_palette_few}},
  \code{\link{distinct_palette_many}},
  \code{\link{gradient_palette}},
  \code{\link{plot_DET_index}},
  \code{\link{plot_drivers_clonality}},
  \code{\link{plot_drivers_graph}},
  \code{\link{plot_drivers_occurrence}},
  \code{\link{plot_jackknife_cluster_stability}},
  \code{\link{plot_jackknife_coclustering}},
  \code{\link{plot_patient_CCF_histogram}},
  \code{\link{plot_patient_data}},
  \code{\link{plot_patient_mutation_burden}},
  \code{\link{plot_patient_oncoprint}},
  \code{\link{plot_patient_trees_scores}}
}
\concept{Plotting functions}
