% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patient_data.R
\name{plot_patient_data}
\alias{plot_patient_data}
\title{Plot the data for a patient.}
\usage{
plot_patient_data(x, patient, ...)
}
\arguments{
\item{x}{A \code{REVOLVER} cohort.}

\item{patient}{A patient id.}

\item{...}{Extra parameters passed to all the used plotting functions.}
}
\value{
A figure assembled with \code{ggpubr}.
}
\description{
This function creates a complex plot for the
data of a patient, assembling plots returned from the following
functions: 1) \code{\link{plot_data_clusters}},
2) \code{\link{plot_data_clone_size}}, 3) \code{\link{plot_data_mutation_burden}}
and 4) \code{plot_patient_oncoprint}.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

plot_patient_data(TRACERx_NEJM_2017_REVOLVER, 'CRUK0001')
plot_patient_data(TRACERx_NEJM_2017_REVOLVER, 'CRUK0002')
}
\seealso{
Other Plotting functions: \code{\link{distinct_palette_few}},
  \code{\link{distinct_palette_many}},
  \code{\link{gradient_palette}},
  \code{\link{plot_DET_index}},
  \code{\link{plot_drivers_clonality}},
  \code{\link{plot_drivers_graph}},
  \code{\link{plot_drivers_occurrence}},
  \code{\link{plot_patient_CCF_histogram}},
  \code{\link{plot_patient_mutation_burden}},
  \code{\link{plot_patient_oncoprint}},
  \code{\link{plot_patient_trees_scores}}
}
\concept{Plotting functions}
