% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patient_mutation_burden.R
\name{plot_patient_mutation_burden}
\alias{plot_patient_mutation_burden}
\title{Plot the mutation burden for a patient.}
\usage{
plot_patient_mutation_burden(x, patient, ...)
}
\arguments{
\item{x}{A REVOLVER cohort object.}

\item{patient}{A patient id.}

\item{...}{Extra parameters, not used.}

\item{cex}{Cex of the plot.}
}
\value{
A \code{ggplot} plot.
}
\description{
This function creates a scatterplot
of a full cohort annotating clonal versus subclonal
mutational burden, colouring each point by the 
number of total drivers in the patient, and using
as shape the number of clones with drivers.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')
 
plot_patient_mutation_burden(TRACERx_NEJM_2017_REVOLVER, 'CRUK0001')
plot_patient_mutation_burden(TRACERx_NEJM_2017_REVOLVER, 'CRUK0002')
}
\seealso{
Other Plotting functions: \code{\link{distinct_palette_few}},
  \code{\link{distinct_palette_many}},
  \code{\link{gradient_palette}},
  \code{\link{plot_DET_index}},
  \code{\link{plot_drivers_clonality}},
  \code{\link{plot_drivers_graph}},
  \code{\link{plot_drivers_occurrence}},
  \code{\link{plot_patient_CCF_histogram}},
  \code{\link{plot_patient_data}},
  \code{\link{plot_patient_oncoprint}},
  \code{\link{plot_patient_trees_scores}}
}
\concept{Plotting functions}
