% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patient_trees_scores.R
\name{plot_patient_trees_scores}
\alias{plot_patient_trees_scores}
\title{Plot the scores for a patient's trees.}
\usage{
plot_patient_trees_scores(x, patient,
  transfer_palette = distinct_palette_few, ...)
}
\arguments{
\item{x}{A REVOLVER object.}

\item{patient}{The id of a patient.}

\item{transfer_palette}{A function that can be used to sample
an arbitrary number of colours to identify the transfers.}
}
\value{
A \code{ggplot} plot.
}
\description{
Plots a barplot where each tree is
a bar, with height proportional to the tree score and
colour to its particular combination of transfer.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

plot_patient_trees_scores(TRACERx_NEJM_2017_REVOLVER, 'CRUK0001')
plot_patient_trees_scores(TRACERx_NEJM_2017_REVOLVER, 'CRUK0002')
}
\seealso{
Other Plotting functions: \code{\link{distinct_palette_few}},
  \code{\link{distinct_palette_many}},
  \code{\link{gradient_palette}},
  \code{\link{plot_DET_index}},
  \code{\link{plot_drivers_clonality}},
  \code{\link{plot_drivers_graph}},
  \code{\link{plot_drivers_occurrence}},
  \code{\link{plot_patient_CCF_histogram}},
  \code{\link{plot_patient_data}},
  \code{\link{plot_patient_mutation_burden}},
  \code{\link{plot_patient_oncoprint}}
}
\concept{Plotting functions}
