% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_penalty.R
\name{plot_penalty}
\alias{plot_penalty}
\title{Plot the fit penalty}
\usage{
plot_penalty(x, drivers = x$variantIDs.driver, min.occurrences = 0,
  alpha_level = 0.05,
  drivers_palette = colorRampPalette(RColorBrewer::brewer.pal(n = 9,
  "Set1")), cex = 1, ...)
}
\arguments{
\item{x}{A REVOLVER cohort with fit models.}

\item{drivers}{The list of drivers to use; by default all of them. If the
entry is a subset of the actual list of all drivers, all the entries in the
penalty data structure \code{x$fit$penalty} will be used if they involve at
least one event from \code{drivers}.}

\item{min.occurrences}{The penalty data structure will be filtered for
\code{count} values above this threshold.}

\item{alpha_level}{The significance level for the enrichment Fisher test.}

\item{drivers_palette}{A function that can return, for an input number,
a number of colours.}

\item{cex}{The cex of the plot.}

\item{...}{Other unused parameters.}
}
\value{
A ggplot object for this plot.
}
\description{
Plot the fit penalty as a barplot, for each one of a set of desired
driver events, where the bar represents the counts of each trajectory
in the data. This function allows also to filter out entries that have
been seen below a predetermined cutoff, and tests for significance in the
association A -> B via a one-sided Fisher 2x2 test adjusted for the number of
comparison (marginal count of B-ended trajectories). The tests are carried
out by function \code{revolver:::enrichment_test_incoming_edge}, which can
be used to obtain a tidy representation of the tests' results.
}
\examples{
TODO
}
