% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trees.R
\name{plot_trees}
\alias{plot_trees}
\title{Plot REVOLVER trees for a patient}
\usage{
plot_trees(x, patient, rank = 1, data = "trees", ...)
}
\arguments{
\item{x}{A REVOLVER cohort object}

\item{patient}{The patient for which the trees should be plot}

\item{rank}{A vector that specifies which tree should be plot for a
patient. The trees are stored according to the rank, and this is an 
integer vector which will be used to access the trees. By default, only
the top-rank tree is plot (\code{rank = 1}).}

\item{data}{}

\item{...}{Extra parameters which will be passed to all the 
delegate plotting calls  other functions.}
}
\value{
A list of plots.
}
\description{
This function plots one or more trees for a patient, in two
possible formats (icon and non-icon). The icon format is a tiny
representation of a tree with coloured nodes and edges, which
shows only the information transfer associated to the tree. The
non-icon version is the canonical representaion of the full tree,
with nodes coloured by drivers status, node size scaled by the clone
size (per node), and drivers annotated as labels. Switching from
one representation to the other is possible via a parameter
\code{icon}. This function returns a list of trees (which can
be assembled via \code{ggpubr} functions for instance); the number
of trees that one wants to plot is controlled via parameter \code{rank}.
}
