% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trees_scores.R
\name{plot_trees_scores}
\alias{plot_trees_scores}
\title{Title}
\usage{
plot_trees_scores(x, patients = x$patients,
  transfer_palette = distinct_palette_few, cex = 1)
}
\arguments{
\item{x}{}

\item{patients}{}

\item{transfer_palette}{}

\item{cex}{}
}
\description{
Title
}
