% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{print.rev_cohort}
\alias{print.rev_cohort}
\title{Print for class "rev_cohort"}
\usage{
\method{print}{rev_cohort}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{obj of class "rev_cohort"}

\item{digits}{number of output digits}
}
\value{
nothing
}
\description{
Print for class "rev_cohort"
}
\examples{
data(CRC.cohort)
CRC.cohort
}
