% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_cohort_fit.R
\name{print.rev_cohort_fit}
\alias{print.rev_cohort_fit}
\title{Print a REVOLVER cohort object with fits.}
\usage{
\method{print}{rev_cohort_fit}(x, ...)
}
\arguments{
\item{x}{A REVOLVER cohort object from class \code{"rev_cohort_fit"}}

\item{...}{unused arguments from generic}
}
\value{
Nothing
}
\description{
Print a REVOLVER cohort object with fits.
}
\examples{
data(Breast.fit)
Breast.fit
}
