% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_driver.R
\name{remove_driver}
\alias{remove_driver}
\title{Remove a driver event from the cohort.}
\usage{
remove_driver(x, variantID)
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{variantID}{ID of the driver event to remove.}
}
\value{
A modified cohort where the required event is no longer a driver.
}
\description{
Each event is identied through its ID (variantID); 
with this function, you can remove a driver events, which consists
in flagging it as FALSE in the \code{is.Driver} column of the data, and updating
the information transfer. If you have fit the models, however, you should 
re-run the fit after this modification because this modification does not propagate.
}
\examples{
data(CRC.cohort)
head(CRC.cohort$dataset) # Info to use to remove
new.data = revolver_removeDriver(CRC.cohort, "adenoma_1", "APC", "NOTHING", "1")
head(new.data$dataset)
}
