% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_drivers.R
\name{remove_drivers}
\alias{remove_drivers}
\title{Remove a list of driver events from the cohort.}
\usage{
remove_drivers(x, variantID, check = TRUE)
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{variantID}{Id of the driver event to remove.}
}
\value{
A modified cohort where the required events are no longer annotated as drivers.
}
\description{
Each event is identified through its id (\code{variantID});
with this function, you can remove a list of driver events, which consists
in flagging them as \code{FALSE} in the \code{is.Driver} column of the data, and updating
the information transfer. If you have fit the models or clustered the cohort,
you should re-run the analyses after this modification; for this reason,
any previous result from those analyses is cancelled from the returned object.

Notice also that some patients might be removed by this function, because if
they have no longer drivers then they cannot be fit afterwards.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

print(TRACERx_NEJM_2017_REVOLVER)

new_cohort = remove_drivers(TRACERx_NEJM_2017_REVOLVER, "MET")
print(new_cohort)
}
