% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_patient.R
\name{remove_patient}
\alias{remove_patient}
\title{Remove a patient from the cohort.}
\usage{
remove_patient(x, patientID)
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{patientID}{Id of the patient  to remove.}
}
\value{
A modified cohort without the required patient.
}
\description{
Each patient is identied through its id (\code{patientID});
with this function, you can remove a patient, which consists in removing his data
and trees. If you have fit the models or clustered the cohort,
you should re-run the analyses after this modification; for this reason,
any previous result from those analyses is cancelled from the returned object.

Notice also that some patients might be removed by this function, because if
they have only one driver then they cannot be fit after driver removal.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

print(TRACERx_NEJM_2017_REVOLVER)

new_cohort = remove_patient(TRACERx_NEJM_2017_REVOLVER, "CRUK0001")
print(new_cohort)
}
