% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_analyzer.R
\name{revolver_analyzer}
\alias{revolver_analyzer}
\title{Wrapper for a full analysis with REVOLVER

This is convenient wrapper to run all REVOLVER's analysis within one command.
It takes as input a \code{"rev_cohort"} object and performs the following steps:
1) it computes phylogenetic trees (either from binary or CCF data);
2) it performs model fit;
3) if performs clustering.
This function requires as parameters all the parameters that would have been
passed to the repective functions, where you to run the analysis in single steps.
This function saves relevant RData objects for its main steps.}
\usage{
revolver_analyzer(cohort, type = "CCF",
  cohort.name = "REVOLVER-cohort", folder.output = ".",
  do.plots = TRUE, options.trees = list(sspace.cutoff = 10000,
  n.sampling = 5000, store.max = 200, overwrite = FALSE),
  options.fit = list(initial.solution = NA, transitive.orderings = FALSE,
  restarts = 10), options.clustering.withGL = list(transitive.closure =
  FALSE, min.group.size = 3, hc.method = "ward", cutoff.features_annotation
  = 3, split.method = "cutreeHybrid"),
  options.clustering.withoutGL = list(transitive.closure = FALSE,
  min.group.size = 3, hc.method = "ward", cutoff.features_annotation = 3,
  split.method = "cutreeDynamic"))
}
\arguments{
\item{cohort}{A \code{"rev_cohort"} object.}

\item{type}{Type of data: \code{"CCF"} or \code{"binary"}. If \code{NA} it assumes that \code{"cohort"} contains
already the tree models for your patients.}

\item{cohort.name}{Default is \code{"REVOLVER-cohort"}, use this to identify output files that
will have this string as prefix.}

\item{folder.output}{Output folder, default is current one.}

\item{do.plots}{Set it to \code{TRUE} if you want also plots as output.}

\item{options.trees}{List of parameters for building input trees. See \code{\link{revolver_compute_phylogenies}}.
or \code{\link{revolver_compute_CLtrees}}.}

\item{options.fit}{List of parameters for fitting models. See  \code{\link{revolver_fit}}.}

\item{options.clustering.withGL}{List of parameters for clustering with the germline node GL. See \code{\link{revolver_cluster}}.}

\item{options.clustering.withoutGL}{List of parameters for clustering without the germline node GL. See \code{\link{revolver_cluster}}.}
}
\value{
Nothing, as all functions dump to disk several PDF and RData files.
}
\description{
Wrapper for a full analysis with REVOLVER

This is convenient wrapper to run all REVOLVER's analysis within one command.
It takes as input a \code{"rev_cohort"} object and performs the following steps:
1) it computes phylogenetic trees (either from binary or CCF data);
2) it performs model fit;
3) if performs clustering.
This function requires as parameters all the parameters that would have been
passed to the repective functions, where you to run the analysis in single steps.
This function saves relevant RData objects for its main steps.
}
\examples{
data(CRC.cohort)
revolver_analyzer(CRC.cohort, type = 'binary', do.plots = FALSE)

}
