% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_check_cohort.R
\name{revolver_check_cohort}
\alias{revolver_check_cohort}
\title{Check basic inconsistencies in a REVOLVER cohort.}
\usage{
revolver_check_cohort(x, stopOnError = FALSE)
}
\arguments{
\item{x}{An object of class \code{"rev_cohort"}}

\item{stopOnError}{Whether or not it should raise a stop on error.}
}
\value{
Nothing, all relevant information are print to screen
and the stop is raised only if \code{stopOnError = TRUE}.
}
\description{
Check basic inconsistencies in a REVOLVER cohort.
}
\details{
Perform some basic diagnostic of a cohort object.
It will inform of patients without drivers and other
information that can be used to reshape the data before
fitting a model.
}
\examples{
data(CRC.cohort)
revolver_check_cohort(CRC.cohort)
print(CRC.cohort) # calls this anyway
}
