% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_clustering.R
\name{revolver_cluster}
\alias{revolver_cluster}
\title{Compute hierarchical clustering for a REVOLVER cohort with fit models.}
\usage{
revolver_cluster(x, hc.method = "ward", split.method = "cutreeHybrid",
  dendogram.type = "rectangle", cutoff.features_annotation = 3,
  min.group.size = 2, plot.groups = TRUE,
  file = "REVOLVER-clustering.pdf", ...)
}
\arguments{
\item{x}{A 'rev_cohort_fit' object for which the evolutionary distance has been computed.}

\item{hc.method}{Method for hierarchial clustering, see "revolver_infoclustering".}

\item{split.method}{Method to cut the dendogram, see "revolver_infoclustering".}

\item{dendogram.type}{"rectangle", "triangle" or other formats supported by a dendogram object.}

\item{cutoff.features_annotation}{Minimum number of observations to include a certain feature
in the plot. The number is absolute; e.g., with n = 3 we annotate only edges that occurr at least
in 3 patients}

\item{min.group.size}{Minimum group size for dynamicTreeCut functions.}

\item{plot.groups}{Set this to TRUE if you want to plot also the resulting groups and their models.}

\item{file}{Output file}

\item{...}{Extra parameters}
}
\value{
A 'rev_cohort_fit' object mod a new field "cluster" that contains the computation's results.
}
\details{
Compute hierarchical clustering for a REVOLVER cohort with fit models.
}
\examples{
data(CRC.cohort)
fit = revolver_fit(CRC.cohort)
fit = revolver_evo_distance(fit)
fit = revolver_cluster(fit) # dumped also to disk
}
