% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{revolver_cohort}
\alias{revolver_cohort}
\title{Construct an S3 REVOLVER cohort object of class "rev_cohort".}
\usage{
revolver_cohort(dataset, CCF.parser = revolver:::CCF.parser,
  options = list(ONLY.DRIVER = FALSE, MIN.CLUSTER.SIZE = 10),
  annotation = "")
}
\arguments{
\item{dataset}{A dataframe in the specified format (see Online manual).}

\item{CCF.parser}{A function to parse the format for the encoding of CCF
or binary values for each sequenced region. A possible function is available
inside REVOLVER; since it is not exported use :::CCF.parser to retrieve it.}

\item{options}{A list of 2 parameters that should be a boolean value for
ONLY.DRIVER (use only driver SNVs), and MIN.CLUSTER.SIZE, the minimum cluster size.}

\item{annotation}{A string for annotation of this cohort. This will be prompted
in every print for this object.}
}
\value{
An object of class "rev_cohort"
}
\description{
Construct an S3 REVOLVER cohort object of class "rev_cohort".
}
\examples{
data(CRC)
cohort = revolver_cohort(CRC)
}
