% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{revolver_compute_CLtrees}
\alias{revolver_compute_CLtrees}
\title{Compute mutation trees (Chow-Liu trees with Suppes' conditions) from binary data.}
\usage{
revolver_compute_CLtrees(x, patient, precomputed.trees = NULL,
  precomputed.scores = NULL, options = list(sspace.cutoff = 10000,
  n.sampling = 5000, overwrite = FALSE, store.max = 100), verbose = FALSE)
}
\arguments{
\item{x}{An object of class "rev_cohort"}

\item{patient}{A patient in the cohort, for which phylogenies are created}

\item{precomputed.trees}{If a list a of precomputed trees is available, this list should contain their
adjacency matrix. No computation will be carried out in this case.}

\item{options}{If one wants to generate new phylogenies for this patient, there named parameters should be
passed through this list. "sspace.cutoff = 10000" is the state-space cutoff to generate trees in an
exhaustive fashion, or to Montecarlo sample them. If Montecarlo is chosen "sampling = 5000" tells how
many trees are sampled. "overwrite = FALSE" is boolean flag that says if the patient's current phylogenies
should be overwritten, in case they are already available. "store.max = 100" tells how many trees should
be stored.}

\item{verbose}{output type.}

\item{precomputed.trees}{If a list a of precomputed trees is available, this list should contain their
scores. No computation will be carried out in this case.}
}
\value{
a modififed object of class "rev_cohort" with available phylogeneis for "patient".
}
\details{
This is the analogous of "revolver_compute_phylogenies", but for binary data and hence it computes Chow-Liu trees.
}
\examples{
data(CRC.cohort)
new.cohort = revolver_compute_CLtrees(CRC.cohort, patient = 'adenoma_1')
print(new.cohort$phylogenies$adenoma_1)
}
