% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{revolver_deletePatients}
\alias{revolver_deletePatients}
\title{Subset the patients in the cohort to match the input list.}
\usage{
revolver_deletePatients(x, list)
}
\arguments{
\item{list}{A vector of patient IDs to subset the data to.}

\item{cohort}{An object of class "rev_cohort"}
}
\value{
A REVOLVER cohort with only patients in the list "list".
}
\description{
Subset the patients in the cohort to match the input list.
}
\examples{
data(CRC.cohort)
CRC.cohort
revolver_deletePatients(CRC.cohort, "adenoma_2")
}
