% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_DETindex.R
\name{revolver_DETindex}
\alias{revolver_DETindex}
\title{helper function to calculate diversity indices}
\usage{
revolver_DETindex(x, n.boot = 1, table = FALSE, index = "Shannon",
  type = "after.expansion", driver.only = FALSE)
}
\arguments{
\item{x}{revolver output}

\item{n.boot}{number of bootstraps (default is 1)}

\item{table}{return a table of results? (default is FALSE)}

\item{index}{diversity index (Shannon or VA, default is Shannon)}

\item{type}{'before.expansion' or 'after.expansion' (default after)}

\item{driver.only}{use driver events only? (default is FALSE)}
}
\value{

}
\description{
helper function to calculate diversity indices
}
