% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_clustering.R
\name{revolver_evo_distance}
\alias{revolver_evo_distance}
\title{Compute REVOLVER's evolutionary distance h(P_i, P_j).}
\usage{
revolver_evo_distance(x, use.GL = TRUE, transitive.closure = FALSE)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object.}

\item{use.GL}{Whether or not the germline node (GL) should be used to compute the distance.
Default is \code{TRUE}.}

\item{transitive.closure}{Whether or not the transitive closure of the ordering relations
should be computed, or not. Default is \code{FALSE}.}
}
\value{
A \code{"rev_cohort_fit"}. object with a new field \code{"cluster"} that contains the computation's results.
}
\description{
Compute REVOLVER's evolutionary distance h(P_i, P_j).
}
\details{
This function can be applied only to an object that contains a fit result,
which is then of class \code{"rev_cohort_fit"}.
Default values represent the most common configuration that one should use.
}
\examples{
data(CRC.cohort)
fit = revolver_fit(CRC.cohort)
fit =revolver_evo_distance(fit)
}
