% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_clustering.R
\name{revolver_infoclustering}
\alias{revolver_infoclustering}
\title{Compute possible clusterings for standard parameters.}
\usage{
revolver_infoclustering(x, methods = c("average", "single", "complete",
  "ward"), min.group.size = 2, do.plot = FALSE, file = "Rplot.pdf")
}
\arguments{
\item{x}{A 'rev_cohort_fit' object for which the evolutionary distance has been computed.}

\item{methods}{Dendogram methods supported by agnes and cluster packages. Default
are c( "average", "single", "complete", "ward").}

\item{min.group.size}{Minimum group size for dynamicTreeCut functions.}

\item{do.plot}{TRUE if you want plots to be saved to file "file"}

\item{file}{PDF output file.}
}
\value{
None
}
\details{
Compute possible clusterings for standard parameters. This helps
to determine and select clusters in an output dendogram. This function uses
dendogram cutting strategies available through packages such as cluster and
dynamicTreeCut.
}
\examples{
data(CRC.cohort)
fit = revolver_fit(CRC.cohort)
fit = revolver_evo_distance(fit)
revolver_infoclustering(fit) # dumped to disk
}
