% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_jackknife.R
\name{revolver_jackknife}
\alias{revolver_jackknife}
\title{Jackknife estimate}
\usage{
revolver_jackknife(cohort, cohort.name = "REVOLVER-cohort",
  folder.output = ".", do.plots = TRUE, resamples = 100,
  leave.out = 0.1, cores.ratio = 0.9, options.fit = list(initial.solution
  = NA, transitive.orderings = FALSE, restarts = 10),
  options.clustering = list(use.GL = TRUE, transitive.closure = FALSE,
  min.group.size = 3, hc.method = "ward", cutoff.features_annotation = 3,
  split.method = "cutreeHybrid"))
}
\arguments{
\item{options.clustering}{}
}
\description{
Jackknife estimate
}
