% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_fit.R
\name{revolver_penaltyPlot}
\alias{revolver_penaltyPlot}
\title{Plot penalty for a REVOLVER cohort, via non parametric bootstrap.}
\usage{
revolver_penaltyPlot(x, n.boot = 100, DET.type = "Shannon",
  file = "REVOLVER-penalty.pdf", ...)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object}

\item{n.boot}{Number of non-parametric bootstrap resamples.}

\item{DET.type}{Index type, default is \code{"Shannon"} (Shannon's equitability,
which is entropy), also available is \code{"VA"} (Variance Analog index in \code{VA})}

\item{file}{Output file.}

\item{...}{Extra parameters.}
}
\value{
none
}
\description{
Plot penalty for a REVOLVER cohort, via non parametric bootstrap.
}
\examples{
data(CRC.cohort)
fit = revolver_fit(CRC.cohort)
revolver_penaltyPlot(fit)
}
