% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_phylogeny.R
\name{revolver_phylogeny}
\alias{revolver_phylogeny}
\title{REVOLVER constructor for a tree}
\usage{
revolver_phylogeny(x, patient, M, score,
  annotation = paste0("A tree for patient ", patient))
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{patient}{The id of this patient in the cohort.}

\item{M}{The tree adjacency matrix.}

\item{score}{A score for this model, that we seek to maximize.}

\item{annotation}{A string to annotate this tree.}
}
\value{
An object of class \code{"rev_phylo"} that represents this tree.
}
\description{
REVOLVER constructor for a tree
}
\details{
This constructor creates an object of type "rev_phylo", which represents a model
for a patient. This is the same type of object regarldess data are CCF or binary values.
}
\examples{
# Take data from CRC, inspect sample and dataset
data("CRC")
dataset = CRC[CRC$patientID == 'adenoma_3', ]
samples = paste('R', 1:5, sep = '')

# Extract CCF values
CCF = sapply(dataset$CCF, revolver:::CCF.parser)
CCF = t(apply(CCF, 2, as.numeric))
rownames(CCF) = rownames(dataset)
colnames(CCF) = samples

# Bind a dataset with explicit data
dataset = cbind(dataset, CCF)

# Create empty adj_mat
m = matrix(0, ncol = 4, nrow = 4)
colnames(m) = rownames(m) = 1:4
# Phylogeny: will attach a GL (germline) to the root
# of the tree. In this case all nodes are roots.
revolver_phylogeny(m, 'adenoma_1', dataset, samples, 52)
}
