% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_jackknife.R
\name{revolver_plot_jackknife}
\alias{revolver_plot_jackknife}
\title{Plot jackknife statistics.}
\usage{
revolver_plot_jackknife(x, sorting = "heterogeneity",
  cutoff.annotate.numbers = 0.6, file = "REVOLVER-jackknife.pdf",
  jam.PDFs = TRUE, ...)
}
\arguments{
\item{x}{An object obtained as output of \code{\link{revolver_jackknife}}.}

\item{sorting}{For edges, default sorting is "heterogeneity", which means that we count how many times we find a
certain edge (regardless its value estimated via the jackknife). Any different value will render the
ordering alphabetical.}

\item{file}{Output file, because multiple plots are generated this should always be a file.}

\item{...}{Parameters forwarded to the \code{pheatmap} call}
}
\value{
Nothing.
}
\description{
Plot jackknife statistics.
}
\details{
It plots a heatmap with the result of the jackknife computations.
}
\examples{
TODO
}
