% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_clustering.R
\name{revolver_plt_compare_dendrograms}
\alias{revolver_plt_compare_dendrograms}
\title{Plot the comparison among REVOLVER's clusters and the one obtained
by analyzing the occurence of mutations.}
\usage{
revolver_plt_compare_dendrograms(x, cex = 1, type = "binary",
  dendrogram.type = "rectangle", file = NA)
}
\arguments{
\item{x}{A cohort object with fits and clusters.}

\item{cex}{Cex for the plot}

\item{type}{Compare against "binary" or "clonality" (see Description).}

\item{dendrogram.type}{Type of dendrogram ("rectangle", "triangle", "circular")}

\item{file}{Output file, if NA no file is used}
}
\value{
nothing
}
\description{
Plot the comparison among REVOLVER's clusters and the one obtained
by analyzing the occurence of mutations.
}
\details{
This function can be applied only to an object that contains a fit result
and clusters. The function will plot the dendrogram of the alternative
clustering method, which will cluster either binary (type = 'binary') or
clonal/ subclonal patterns of occurrence (type = 'clonality'). Besides the
dendrogram, a tanglegram against REVOLVER's clusters is also computed.
}
\examples{
\dontrun{
 TODO
}

}
