% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_clustering.R
\name{revolver_plt_dendogram}
\alias{revolver_plt_dendogram}
\title{Plot the dendrogram for REVOLVER's clusters.}
\usage{
revolver_plt_dendogram(x, type = "REVOLVER", cex = 1,
  dendogram.type = "rectangle", file = NA)
}
\arguments{
\item{x}{A cohort object with fits and clusters.}

\item{type}{What clustering should be used: "REVOLVER" is for REVOLVER clustering,
"binary" or "clonality" are for occurrence-based clusterings. In any case leaves are
colored according to REVOLVER's clusters.}

\item{cex}{Cex for the plot}

\item{dendogram.type}{Type of dendrogram (rectangle, triangle, circular)}

\item{file}{Output file, if NA no file is used}
}
\value{
nothing
}
\description{
Plot the dendrogram for REVOLVER's clusters.
}
\details{
This function can be applied only to an object that contains a fit result
and clusters. The function will plot the dendrogram for the results of
hierarchical clustering, and will color labels by group assignment.
}
\examples{
\dontrun{
 TODO
}

}
