% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_fit.R
\name{revolver_plt_DET_index}
\alias{revolver_plt_DET_index}
\title{Plot the DET index, via non parametric bootstrap.}
\usage{
revolver_plt_DET_index(x, DET.type = "Shannon",
  type = "after.expansion", N = 100, colour = "steelblue",
  file = NA, cex = 1)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object}

\item{DET.type}{Index type, default is \code{"Shannon"} (Shannon's equitability,
which is entropy), also available is \code{"VA"} (Variance Analog index in \code{VA})}

\item{type}{Use this to select either 'after.expansion', or 'before.expansion'
estimates of the penalty}

\item{N}{Number of non-parametric bootstrap resamples.}

\item{colour}{Colour of the bars.}

\item{file}{Output file, or NA.}
}
\value{
The DET values
}
\description{
Plot the index of Divergent Evolutionary Trajectories, computed via
a non parametric bootstrap procedure. The current DET is implemented
in two different ways: via Shannon's entropy or the Variance Analog index.
You can also decide if you want to use models computed after  likelihood fit, or
after likelihood fit and expansion.
}
\examples{
data(Breast.fit)
revolver_plt_DET_index(Breast.fit)
}
