% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_fit.R
\name{revolver_plt_DET_index_driver}
\alias{revolver_plt_DET_index_driver}
\title{Plot the DET index of each driver from the data (empirical).}
\usage{
revolver_plt_DET_index_driver(x, DET.type = "Shannon",
  type = "after.expansion", palette = "Blues", file = NA, cex = 1)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object}

\item{DET.type}{Index type, default is \code{"Shannon"} (Shannon's equitability,
which is entropy), also available is \code{"VA"} (Variance Analog index in \code{VA})}

\item{type}{Use this to select either 'after.expansion', or 'before.expansion'
estimates of the penalty}

\item{palette}{RColorBrewer palette}

\item{file}{Output file, or NA.}
}
\value{
The DET values
}
\description{
The same index of \code{\link{revolver_plt_DET_index}} but computed per driver
and without the bootstrap.
}
\examples{
data(Breast.fit)
revolver_plt_DET_index_driver(Breast.fit)
}
