% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_clustering.R
\name{revolver_plt_fit_by_group}
\alias{revolver_plt_fit_by_group}
\title{Plot model fit grouped by cluster.}
\usage{
revolver_plt_fit_by_group(x, cex = 1,
  file.suffix = "REVOLVER-Clusters-FitsPerCluster.pdf")
}
\arguments{
\item{x}{A cohort object with fits and clusters.}

\item{cex}{Cex for the plot.}

\item{file.suggix}{Each filename will be suffixed by this string,
and annotated also with the cluster ID.}
}
\value{
nothing
}
\details{
This function can be applied only to an object that contains a fit result
and clusters. For each cluster, this function plots the fit of the models
in the cluster. All outputs are directed to separate PDF files.
}
\examples{
TODO
}
