% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_fit.R
\name{revolver_plt_fit_patient}
\alias{revolver_plt_fit_patient}
\title{Plot the tree fit for a patient.}
\usage{
revolver_plt_fit_patient(x, patient, palette = "Set1", alpha = 0.8,
  file = NA, cex = 1)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object}

\item{patient}{The patient to plot.}

\item{palette}{RColorBrewer palette.}

\item{alpha}{Alpha colouring.}

\item{file}{Output file, or NA.}

\item{cex}{Cex for graphics.}
}
\value{
None
}
\description{
This function uses REVOLVER's standard plotting functions but shows the fit
for a model.
}
\examples{
data(Breast.fit)
revolver_plt_fit_patient(Breast.fit, patient = 'PD9850')
}
