% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_clustering.R
\name{revolver_plt_group_trajectories}
\alias{revolver_plt_group_trajectories}
\title{Plot the consensus model of the trajectories/}
\usage{
revolver_plt_group_trajectories(x, cex = 1, cutoff.edges_annotation = 3,
  file = NA, jamPDF = FALSE, jam.layout = "2x1")
}
\arguments{
\item{x}{A cohort object with fits and clusters.}

\item{cex}{Cex for the plot.}

\item{cutoff.edges_annotation}{Edges whose frequency in each group is below
this threshold will not be plot.}

\item{file}{Output file, if NA no file is used}

\item{jamPDF}{If TRUE, the produced figures will be jammed. This makes sense
only if you set file not NA.}

\item{jam.layout}{Layout to jam PDFS, e.g. '2x1'}
}
\value{
nothing
}
\details{
This function can be applied only to an object that contains a fit result
and clusters. For each group and for the overall cohort, this function
plots the consensus model of the edges, annotated with counts.
}
\examples{
\dontrun{
 TODO
}

}
