% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_fit.R
\name{revolver_plt_itransfer_patient}
\alias{revolver_plt_itransfer_patient}
\title{Plot the information transfer from a patient's fit.}
\usage{
revolver_plt_itransfer_patient(x, patient, file = NA, cex = 1)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object}

\item{patient}{The patients to plot.}

\item{file}{Output file, or NA.}

\item{cex}{Cex for graphics.}
}
\value{
None
}
\description{
Plot the information transfer from a patient's fit.
}
\details{
From the phylogenetic tree fit to this patient, REVOLVER extracts and plots the
information transfer depicted by the tree. The tree is annotated with the number
of clonal and subclonal observations of each driver, and the amount of times each
edge is detected
}
\examples{
data(Breast.fit)
revolver_plt_itransfer_patient(Breast.fit, patient = 'PD9850')
}
