% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_jackknife.R
\name{revolver_plt_jackknife_coclust}
\alias{revolver_plt_jackknife_coclust}
\title{Plot jackknife statistics for co-clustering (heatmap).}
\usage{
revolver_plt_jackknife_coclust(x, cutoff.annotate.numbers = 0.6,
  palette = "YlGnBu", file = NA, ...)
}
\arguments{
\item{x}{An object obtained as output of \code{\link{revolver_jackknife}}.}

\item{cutoff.annotate.numbers}{Numbers above this value will also be written in the plot.}

\item{palette}{RColorBrewer palette}

\item{file}{Output file, or NA.}

\item{...}{Parameters forwarded to the \code{pheatmap} call}
}
\value{
None
}
\description{
Plot jackknife statistics for co-clustering (heatmap).
}
\details{
It plots a heatmap with the result of the jackknife computations
for the co-clustering probability.
}
\examples{
data(Breast.fit)
revolver_plt_jackknife_coclust(Breast.fit)
}
