% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_jackknife.R
\name{revolver_plt_jackknife_edge_counts}
\alias{revolver_plt_jackknife_edge_counts}
\title{Plot jackknife statistics for the number of patients with an edge.}
\usage{
revolver_plt_jackknife_edge_counts(x, cutoff.annotate.numEdges = 3, cex = 1,
  file = NA)
}
\arguments{
\item{x}{An object obtained as output of \code{\link{revolver_jackknife}}.}

\item{cutoff.annotate.numEdges}{Minimum mean value to filter some low-freq edges.}

\item{cex}{Cex graphics.}

\item{file}{Output file, because multiple plots are generated this should always be a file.}
}
\value{
None
}
\description{
Plot jackknife statistics for the number of patients with an edge.
}
\details{
It plots a boxplot with the result of the jackknife computations.
}
\examples{
data(Breast.fit)
revolver_plt_jackknife_edge_counts(Breast.fit)
}
