% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_jackknife.R
\name{revolver_plt_jackknife_edge_prb}
\alias{revolver_plt_jackknife_edge_prb}
\title{Plot jackknife statistics for edge detection probability.}
\usage{
revolver_plt_jackknife_edge_prb(x, sorting = "heterogeneity",
  cutoff.annotate.numbers = 0.6, palette = "YlGnBu", file = NA, ...)
}
\arguments{
\item{x}{An object obtained as output of \code{\link{revolver_jackknife}}.}

\item{sorting}{Default is "heterogeneity", which means that we count how many times we find a
certain edge (regardless its value estimated via the jackknife). Any different value will render the
ordering alphabetical.}

\item{palette}{RColorBrewer palette}

\item{file}{Output file, because multiple plots are generated this should always be a file.}

\item{...}{Parameters forwarded to the \code{pheatmap} call}
}
\value{
Nothing.
}
\description{
Plot jackknife statistics for edge detection probability.
}
\details{
It plots a heatmap with the result of the jackknife computations.
}
\examples{
data(Breast.fit)
revolver_plt_jackknife_edge_prb(Breast.fit)
}
