% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_cohort.R
\name{revolver_plt_patient_data}
\alias{revolver_plt_patient_data}
\title{Plot a patient's data}
\usage{
revolver_plt_patient_data(x, patient, cex = 1, file = NA)
}
\arguments{
\item{x}{An object of class \code{"rev_cohort"}}

\item{patient}{The patient to plot}

\item{cex}{Cex for graphics (pheatmap cells)}

\item{file}{Output file, if not NA.}
}
\value{
nothing
}
\description{
Plot a patient's data
}
\examples{
data(Breast.fit)
revolver_plt_patient_data(Breast.fit, 'PD14767')
}
