% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_cohort.R
\name{revolver_plt_patient_trees}
\alias{revolver_plt_patient_trees}
\title{Plot a the top-scoring trees for a patient.}
\usage{
revolver_plt_patient_trees(x, patient, max.phylogenies = 12, file = NA,
  cex = 1)
}
\arguments{
\item{x}{An object of class \code{"rev_cohort"}}

\item{patient}{The patient to plot.}

\item{max.phylogenies}{How many trees should be computed.}

\item{file}{Output file, or NA.}

\item{cex}{Scale cex for graphics.}
}
\value{
nothing
}
\description{
Plot a the top-scoring trees for a patient.
}
\details{
Iterative plotting for the set of input trees of a patient (mutation
trees or phylogenetic trees). This is usually used when ispecting the
information transfer for the input patients.
}
\examples{
data(Breast.fit)
revolver_plt_patient_trees(Breast.fit, 'PD14767')
}
