% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_fit.R
\name{revolver_plt_penalty_barplot}
\alias{revolver_plt_penalty_barplot}
\title{Plot a barplot of the penalty per patient.}
\usage{
revolver_plt_penalty_barplot(x, file = NA, cex = 1)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object}

\item{file}{Output file, leave it to NA to plot to canvas.}
}
\value{
none
}
\description{
Plot a barplot of the penalty per patient.
}
\examples{
data(Breast.fit)
revolver_plt_penalty_barplot(Breast.fit)
}
