% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_fit.R
\name{revolver_plt_penalty_matrix}
\alias{revolver_plt_penalty_matrix}
\title{Plot a pheatmap of the penalty (patients per alterations) .}
\usage{
revolver_plt_penalty_matrix(x, palette = "Blues", normalized = FALSE,
  type = "after.expansion", file = NA, cex = 1)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object}

\item{palette}{RColorBrewer palette}

\item{normalized}{Set to TRUE to normalize the penalty.}

\item{type}{Use this to select either 'after.expansion', or 'before.expansion'
estimates of the penalty}

\item{file}{Output file, leave it to NA to plot to canvas.}
}
\value{
none
}
\description{
This function plots the multinomial model "w". It can plot its
value after the first step of the fitting algorithm, or the second.
It plots normalized or unnormalized values.
}
\examples{
data(Breast.fit)
revolver_plt_penalty_matrix(Breast.fit)
}
