% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_clustering.R
\name{revolver_plt_rbannerplot}
\alias{revolver_plt_rbannerplot}
\title{Bannerplot for REVOLVER's clusters.}
\usage{
revolver_plt_rbannerplot(x, cex = 1, file = NA)
}
\arguments{
\item{x}{A cohort object with fits and clusters.}

\item{cex}{Cex for the plot}

\item{file}{Output file, if NA no file is used}
}
\value{
nothing
}
\description{
Bannerplot for REVOLVER's clusters.
}
\details{
This function can be applied only to an object that contains a fit result
and clusters. The function will plot the bannerplot for the results of
hierarchical clustering.
}
\examples{
\dontrun{
 TODO
}

}
