% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_clustering.R
\name{revolver_plt_rclusters}
\alias{revolver_plt_rclusters}
\title{Plot the features of REVOLVER's clusters, in a heatmap.}
\usage{
revolver_plt_rclusters(x, cex = 1, cutoff.features_annotation = 2,
  file = NA)
}
\arguments{
\item{x}{A cohort object with fits and clusters.}

\item{cex}{Cex for the plot}

\item{cutoff.features_annotation}{Minimum number of occurrences that a feature needs to have
to include it in the heatmap.}

\item{file}{Output file, if NA no file is used}
}
\value{
nothing
}
\details{
This function can be applied only to an object that contains a fit result
and clusters. The function will plot the heatmap of all features that define
REVOLVER's clusters. These are the data, as well as the repeated evolutionary
trajectories detected by the method. This plot is probably the most common one
that you want to inspect to analyze your clusters
}
\examples{
TODO
}
