% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_fit.R
\name{revolver_plt_trajectories_patient}
\alias{revolver_plt_trajectories_patient}
\title{Plot the trajectories in a patient's fit model.}
\usage{
revolver_plt_trajectories_patient(x, patient, palette = "Set1", alpha = 0.8,
  file = NA, cex = 1)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object}

\item{patient}{The patient to plot.}

\item{palette}{RColorBrewer palette.}

\item{alpha}{Alpha colouring.}

\item{file}{Output file, or NA.}

\item{cex}{Cex for graphics.}
}
\value{
None
}
\description{
From the phylogenetic tree fit to this patient, REVOLVER extracts and plots the
trajectories depicted by the tree.
}
\examples{
data(Breast.fit)
revolver_plt_trajectories_patient(Breast.fit, patient = 'PD9850')
}
