% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{revolver_removeDriver}
\alias{revolver_removeDriver}
\title{Remove a driver event from a cohort.}
\usage{
revolver_removeDriver(x, patient, variantID, misc, cluster)
}
\arguments{
\item{x}{An object of class \code{"rev_cohort"}}

\item{patient}{Patient ID.}

\item{variantID}{ID of the event to mark as \code{FALSE}.}

\item{misc}{misc of the event to mark as \code{FALSE}}

\item{cluster}{cluster of the event to mark as \code{FALSE}}
}
\value{
a modified cohort object of class \code{"rev_cohort"}
}
\description{
Remove a driver event from a cohort.
}
\details{
Basic editing function. Each event is identied through its ID,
cluster assignment and its misc flag. With this function, you can remove it.
The event is not physically removed from the dataset, but instead its flag
\code{is.Driver} is set to \code{FALSE}. If there are phylogenies inside, they are updated
as well. If you have fit the models, however, you should re-run the fit after this
modification because this modification does not propagate.
}
\examples{
data(CRC.cohort)
head(CRC.cohort$dataset) # Info to use to remove
new.data = revolver_removeDriver(CRC.cohort, "adenoma_1", "APC", "NOTHING", "1")
head(new.data$dataset)
}
