% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_report_fit_patient.R
\name{revolver_report_fit_patient}
\alias{revolver_report_fit_patient}
\title{Generate PDF report for a patient's fit}
\usage{
revolver_report_fit_patient(x, patient, cex = 1,
  file = paste0("REVOLVER-report-fit-patient-", patient, ".pdf"))
}
\arguments{
\item{x}{An object of class \code{"rev_cohort_fit"}}

\item{patient}{The patient to plot}

\item{cex}{Cex for graphics}

\item{file}{Output file, cannot be NA.}
}
\description{
Generate PDF report for a patient's fit
}
\examples{
data(Breast.fit)
revolver_report_fit_patient(Breast.fit, 'PD9770')

}
