% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_report.R
\name{revolver_report_patient}
\alias{revolver_report_patient}
\title{Generate PDF report for a patient's data and trees}
\usage{
revolver_report_patient(x, patient, cex = 1,
  file = paste0("REVOLVER-report-patient-", patient, ".pdf"))
}
\arguments{
\item{x}{An object of class \code{"rev_cohort"}}

\item{patient}{The patient to plot}

\item{cex}{Cex for graphics}

\item{file}{Output file, cannot be NA.}
}
\value{
nothing
}
\examples{
data(Breast.fit)
revolver_report_patient(Breast.fit, 'PD9770')
}
