% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{revolver_subsetDrivers}
\alias{revolver_subsetDrivers}
\title{Subset the drivers in the cohort to match the input list.}
\usage{
revolver_subsetDrivers(cohort, list)
}
\arguments{
\item{cohort}{An object of class "rev_cohort"}

\item{list}{A vector of driver IDs to subset the data to.}
}
\value{
A REVOLVER cohort with only drivers in the list "list".
}
\description{
Subset the drivers in the cohort to match the input list.
}
\examples{
TODO
}
