% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Stats}
\alias{Stats}
\title{Return summary stastics for the cohort's patients}
\usage{
Stats(x, patients = x$patients)
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{patients}{The patients for which the summaries are computed.}
}
\value{
A tibble with summary stastics.
}
\description{
Returns the number of samples per patient, the number
of drivers, the number of clonal and subclonal mutations etc. The
function can be run on a subset of patients.
}
\examples{
data(CRC.cohort)
Stats(CRC.cohort)
}
