% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Stats_cohort}
\alias{Stats_cohort}
\title{See \code{Stats} function.}
\usage{
Stats_cohort(...)
}
\arguments{
\item{...}{Parameters forwarded to the \code{Stats} function.}
}
\value{
A tibble with summary stastics.
}
\description{
Just a wrapper to the \code{Stats} function.
}
\examples{
data(CRC.cohort)
Stats_cohort(CRC.cohort)
}
