% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Stats_drivers}
\alias{Stats_drivers}
\title{Return summary stastics for the cohort's drivers}
\usage{
Stats_drivers(x, drivers = x$variantIDs.driver)
}
\arguments{
\item{x}{a REVOLVER object of class "rev_cohort"}

\item{drivers}{The drivers for which the summaries are computed.}
}
\value{
A tibble with the driver stastics.
}
\description{
Returns the number of clonal and subclonal occurrences
of the drivers in the cohort, and their percentage relative to the
cohort size. The function can be run on a subset of drivers.
}
\examples{
data(TRACERx_cohort)

# Get the stats for all patients
Stats_drivers(TRACERx_cohort) 

# And subset the patients
Stats_drivers(TRACERx_cohort, patients = c('CRUK0001', 'CRUK0002')) 
}
