% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{Stats_fits}
\alias{Stats_fits}
\title{Summary statistics for the cohort's fits.}
\usage{
Stats_fits(x, patients = x$patients)
}
\arguments{
\item{x}{A REVOLVER cohort where fits have been computed.}

\item{patients}{The patients for which the summaries are required.}
}
\value{
A tibble with the fits stastics.
}
\description{
Returns a tibble that extends the result of 
\code{\link{Stats_trees}} with information about the fit models.
Compared to summaries returns by other \code{Stats_*} functions,
the information from this one is precomputed.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

# Get the stats for all patients
Stats_fits(TRACERx_NEJM_2017_REVOLVER) 

# And subset the patients
Stats_fits(TRACERx_NEJM_2017_REVOLVER, patients = c('CRUK0001', 'CRUK0002')) 
}
\seealso{
Other Summary statistics: \code{\link{Stats_cohort}},
  \code{\link{Stats_drivers}}, \code{\link{Stats_trees}},
  \code{\link{Stats}}
}
\concept{Summary statistics}
