% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Stats_trees}
\alias{Stats_trees}
\title{Return summary stastics for the cohort's trees}
\usage{
Stats_trees(x, patients = x$patients)
}
\arguments{
\item{x}{a REVOLVER object of class "rev_cohort"}

\item{patients}{The patients for which the summaries are computed.}
}
\value{
A tibble with the driver stastics.
}
\description{
Returns the number of clonal and subclonal occurrences
of each driver in the cohort, and their percentage relative to the
cohort size. The function can be run on a subset of patients.
}
\examples{
data(TRACERx_cohort)

# Get the stats for all patients
Stats_trees(TRACERx_cohort) 

# And subset the patients
Stats_trees(TRACERx_cohort, patients = c('CRUK0001', 'CRUK0002')) 
}
