% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Subclonal}
\alias{Subclonal}
\title{Extract CCF/binary data for subclonal mutations of a patient}
\usage{
Subclonal(x, p)
}
\arguments{
\item{x}{a REVOLVER object of class "rev_cohort" or "rev_cohort_fit"}

\item{p}{a patient}
}
\value{
CCF/binary data for \code{p}'s subclonal mutations
}
\description{
Extract CCF/binary data for subclonal mutations of a patient
}
\examples{
data(CRC.cohort)
Subclonal(CRC.cohort, 'adenoma_3')
}
