% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_phylogeny.R
\name{summary.rev_phylo}
\alias{summary.rev_phylo}
\title{Summary of a \code{"rev_phylo"} object.}
\usage{
\method{summary}{rev_phylo}(x, ..., print.stat = TRUE)
}
\arguments{
\item{x}{An object of class \code{"rev_phylo"}.}

\item{...}{Extra parameters}

\item{print.stat}{Print or not statistics to screen.}
}
\value{
none
}
\details{
Reports some summary for a \code{"rev_phylo"} object, this is a bit more than just using print.
}
\examples{
data(CRC.cohort)
summary(CRC.cohort$phylogenies[['adenoma_3']][[1]])
}
