% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRACERx_cohort}
\alias{TRACERx_cohort}
\title{Analysis of the TRACERx cohort, with REVOLVER.}
\format{A REVOLVER cohort object with fits, etc.}
\usage{
data(TRACERx_cohort)
}
\description{
The input object for the analysis is available via the external package
}
\examples{
data(TRACERx_cohort)

TRACERx_cohort
}
\keyword{datasets}
