% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Truncal}
\alias{Truncal}
\title{Extract truncal mutations in a patient}
\usage{
Truncal(x, p)
}
\arguments{
\item{x}{A REVOLVER object of class "rev_cohort" or "rev_cohort_fit"}

\item{p}{A patient}
}
\value{
CCF/binary data for \code{p}'s truncal mutations
}
\description{
Extract truncal mutations in a patient
}
\examples{
data(CRC.cohort)
Drivers(CRC.cohort, 'adenoma_3')
}
