% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{CCF_clusters}
\alias{CCF_clusters}
\title{Extract all clusters in the data of a patient.}
\usage{
CCF_clusters(x, p)
}
\arguments{
\item{x}{A \code{REVOLVER} cohort.}

\item{p}{The id of a patient in the cohort.}
}
\value{
CCF or binary clusters data for a custom patient.
}
\description{
From a cohort object, this function extracts a tibble of
all clusters in the data of a patient.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

CCF_clusters(TRACERx_NEJM_2017_REVOLVER, 'CRUK0002')

CCF_clusters(TRACERx_NEJM_2017_REVOLVER, 'CRUK0008')
}
\seealso{
Other Getters: 
\code{\link{CCF}()},
\code{\link{Clonal_cluster}()},
\code{\link{Data}()},
\code{\link{Drivers}()},
\code{\link{ITransfer}()},
\code{\link{Phylo}()},
\code{\link{Samples}()},
\code{\link{Subclonal}()},
\code{\link{Truncal}()},
\code{\link{get_features}()}
}
\concept{Getters}
