% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Clonal_cluster}
\alias{Clonal_cluster}
\title{Return the id of the clonal cluster for this patient}
\usage{
Clonal_cluster(x, p)
}
\arguments{
\item{x}{a REVOLVER object of class "rev_cohort"}

\item{p}{a patient}
}
\value{
The id of the clonal cluster for \code{p}
}
\description{
Return the id of the clonal cluster for this patient
}
\examples{
data(CRC.cohort)
Clonal_cluster(CRC.cohort, 'adenoma_3')
}
