% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Cluster}
\alias{Cluster}
\title{Get REVOLVER clusters information for this cohort}
\usage{
Cluster(x, patients = x$patients)
}
\arguments{
\item{x}{A REVOLVER cohort with fits and clusters.}

\item{patients}{Patients to use, all by default.}
}
\value{
A tibble with the required clustering assignments
}
\description{
For a set of `patients` - all by default - extract the clustering information
computed by REVOVLER. The result is just a tibble with a column representing
the patient id, and a column with the cluster label.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

Cluster(TRACERx_NEJM_2017_REVOLVER)

#' Cluster(TRACERx_NEJM_2017_REVOLVER, patients = c("CRUK0001", "CRUK0002"))
}
