% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mutation_trees.R
\name{compute_mutation_trees}
\alias{compute_mutation_trees}
\title{Compute binary mutation trees for a patient (phylogenies).}
\usage{
compute_mutation_trees(x, patients = Stats_cohort(x) \%>\%
  pull(patientID), overwrite = FALSE, ...)
}
\arguments{
\item{x}{A \code{REVOLVER} cohort.}

\item{patients}{A set of patient ids in the cohort, for which the
phylogenies are created.}

\item{x}{A \code{REVOLVER} cohort with now available
phylogeneis for the required \code{patients}.}
}
\description{
Create mutation tree phylogenies from the package
\code{btree}, fitting the input data given to \code{REVOLVER}.
A set of patient ids can be given as input, by default all are
used. A parmeter controls if you want to overwrite the trees
of the patient, where they already computed. Please refer to the
parameters of the \code{btree} constructor from package \code{btree}
for the input parameters that you can use to tune the construction.
}
\note{
This function is not yet implemented
}
\examples{
print('TODO')
}
\seealso{
Other Cohort creation: \code{\link{CCF_parser}},
  \code{\link{compute_clone_trees}},
  \code{\link{input_custom_trees}},
  \code{\link{revolver_check_cohort}},
  \code{\link{revolver_cohort}}
}
\concept{Cohort creation}
