% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRC.R, R/CRC.cohort.R
\docType{data}
\name{CRC}
\alias{CRC}
\title{Colorectal Cancer Data by Cross et al.}
\format{A dataframe that represents the CRC data analyzed in REVOLVER.
See also the vignette associated to the case study.}
\usage{
data(CRC)

data(CRC.cohort)
}
\description{
Driver mutations detected by Cross et al. in a cohort of colorecta adenomas and carcinomas.

Driver mutations detected by Cross et al. in a cohort of colorecta adenomas and carcinomas.
This dataset has already been wrapped to a REVOLVER cohort object
}
\examples{
data(CRC)
head(CRC)

data(CRC.cohort)
print(CRC.cohort)

}
\references{
Cross et al. (2018), to appear in Nature Ecology and Evolution.

Cross et al. (2018) to appear in Nature Ecology and Evolution.
}
\keyword{datasets}
