% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{distinct_palette_few}
\alias{distinct_palette_few}
\title{Palette of few distinct colours.}
\usage{
distinct_palette_few()
}
\value{
A function that applied to a number returns the required colours.
}
\description{
ColorRamp palette extrapolating \code{Set1} palette of \code{RColorBrewer} package. The original \code{Set1} palette is
sampled for 9 colours.
}
\examples{
distinct_palette_few(2)
}
