% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{distinct_palette_many}
\alias{distinct_palette_many}
\title{Palette of \code{n} distinct colours.}
\usage{
distinct_palette_many(n)
}
\arguments{
\item{n}{The number of required colours.}
}
\value{
A function that applied to a number returns the required colours.
}
\description{
ColorRamp palette extrapolating all qualitative (\code{qual}) palettes from
the \code{RColorBrewer} package. This function requires the number of required colours
as input parameter.
}
\examples{
distinct_palette_many(2)
}
