% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{distinct_palette_many}
\alias{distinct_palette_many}
\title{Palette of \code{n} distinct colours.}
\usage{
distinct_palette_many(n = NULL)
}
\arguments{
\item{n}{The number of required colours. By default this is NULL.}
}
\value{
Either a function that applied to a number returns the required colours,
or the actual sampling function.
}
\description{
ColorRamp palette extrapolating all qualitative (\code{qual}) palettes from
the \code{RColorBrewer} package. This function can have the number of required colours
as input parameter; if this does not happen a function is returned.
}
\examples{
distinct_palette_many(2)
}
\seealso{
Other Plotting functions: \code{\link{distinct_palette_few}},
  \code{\link{gradient_palette}},
  \code{\link{plot_CCF_histogram}},
  \code{\link{plot_DET_index}},
  \code{\link{plot_drivers_clonality}},
  \code{\link{plot_drivers_occurrence}},
  \code{\link{plot_graph_drivers}},
  \code{\link{plot_patient_data}},
  \code{\link{plot_patient_mutation_burden}},
  \code{\link{plot_patient_oncoprint}},
  \code{\link{plot_trees_scores}}
}
\concept{Plotting functions}
