% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_features.R
\name{get_features}
\alias{get_features}
\title{Return summary features for the cohort.}
\usage{
get_features(x, patients = x$patients)
}
\arguments{
\item{x}{}

\item{patients}{}
}
\value{
A set of matrices that report summary features of the models.
}
\description{
Computes a set of matrix-like summary features for the cohort, in the
form of a matrix. It will return the matrix of driver events with mean
CCF, their occurrence across all patients, and at the clonal and subclonal
level, and the occurrence of all evolutionary trajectories across patients
}
\examples{
TODO
}
