% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{gradient_palette}
\alias{gradient_palette}
\title{Gradient palette.}
\usage{
gradient_palette(n = NULL)
}
\arguments{
\item{n}{The number of required colours. By default this is NULL.}
}
\value{
Either a function that applied to a number returns the required colours,
or the actual sampling function.
}
\description{
ColorRamp palette extrapolating \code{YlGnBu} palette of
\code{RColorBrewer} package. The original \code{YlGnBu} palette is
sampled for 9 colours.
}
\examples{
gradient_palette(2)
}
\seealso{
Other Plotting functions: \code{\link{distinct_palette_few}},
  \code{\link{distinct_palette_many}},
  \code{\link{plot_DET_index}},
  \code{\link{plot_drivers_clonality}},
  \code{\link{plot_drivers_graph}},
  \code{\link{plot_drivers_occurrence}},
  \code{\link{plot_jackknife_cluster_stability}},
  \code{\link{plot_jackknife_coclustering}},
  \code{\link{plot_jackknife_trajectories_stability}},
  \code{\link{plot_patient_CCF_histogram}},
  \code{\link{plot_patient_data}},
  \code{\link{plot_patient_mutation_burden}},
  \code{\link{plot_patient_oncoprint}},
  \code{\link{plot_patient_trees_scores}}
}
\concept{Plotting functions}
