% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_custom_trees.R
\name{input_custom_trees}
\alias{input_custom_trees}
\title{Add a custom set of trees for a patient.}
\usage{
input_custom_trees(x, patient, trees, scores)
}
\arguments{
\item{x}{A REVOLVER cohort object.}

\item{patient}{A patient in the cohort, for which the trees are created.}

\item{trees}{A list a of precomputed adjacency matrices to describe the input trees.}

\item{scores}{A vectore of scores for the input trees.}
}
\value{
A modififed REVOLVER cohort with available phylogeneis for \code{patient}.
}
\description{
For a patient it is possible to add a set of precomputed trees, which need
to be stored as adjacency matrices, and must be scored and ranked by scoring.
The trees must have as nodes all the clusters that are annotated in the data
of the patient. If the patient has already some trees, these are overwritten.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

# To make it simple we use some trees that are already available
trees = Phylo(TRACERx_NEJM_2017_REVOLVER, p = "CRUK0001")

# Get matrices from these objects, and remove the GL columns/ rows
matrices = lapply(trees, function(x) x$adj_mat[rownames(x$adj_mat) != 'GL', colnames(x$adj_mat) != 'GL'])

print(matrices)

# Get scores for these matrices - vector format
scores = sapply(trees, function(x) x$score)

print(scores)

# Add these trees - this function overwrites the original ones
TRACERx_NEJM_2017_REVOLVER = input_custom_trees(
  TRACERx_NEJM_2017_REVOLVER,
  patient = "CRUK0001",
  trees = matrices,
  scores = scores
)
}
\seealso{
Other Cohort creation: \code{\link{CCF_parser}},
  \code{\link{compute_clone_trees}},
  \code{\link{revolver_cohort}}
}
\concept{Cohort creation}
