% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{ITransfer}
\alias{ITransfer}
\title{Extract the information transfer for a patient's tree}
\usage{
ITransfer(x, p, rank = 1, type = "drivers", data = "trees")
}
\arguments{
\item{x}{A REVOLVER cohort}

\item{p}{A patient}

\item{rank}{The rank of the tree to extract}

\item{type}{Either `clones` or `drivers`.}

\item{data}{Either `trees` or `fit`.}
}
\value{
The information transfer for a patient's tree
}
\description{
Extract the information transfer for a patient's tree
}
\examples{
data(TRACERx_cohort)

# Get the transfer among drivers, top-ranking tree
Phylo(TRACERx_cohort, 'CRUK0002', rank = 1, type = 'drivers')

# Get the transfer among clones, top-ranking tree
Phylo(TRACERx_cohort, 'CRUK0002', rank = 1, type = 'clones')

# Get the transfer from the fit
Phylo(TRACERx_cohort, 'CRUK0002', rank = 1, type = 'clones', data = 'fits')
Phylo(TRACERx_cohort, 'CRUK0002', rank = 1, type = 'drivers', data = 'fits')
}
