% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{plot.rev_cohort}
\alias{plot.rev_cohort}
\title{Plotting function for a REVOLVER cohort.}
\usage{
\method{plot}{rev_cohort}(x, patients = x$patients, max.phylogenies = 12,
  cex = 1, plot.stat = TRUE)
}
\arguments{
\item{x}{An object of class "rev_cohort"}

\item{patients}{The patients to plot, default is all the one available.}

\item{max.phylogenies}{How many trees should be computed for each patient}

\item{cex}{Scale cex for graphics}

\item{plot.stat}{Set it to TRUE to visualize also some statistics about each tree}
}
\value{
nothing
}
\details{
Iterative plotting functions that scans some samples from a cohort. It will plot
a patient's data (original and binarized), the score for the trees associated to
the patient, and a number of such trees.
}
\examples{
data(CRC.cohort)
plot(CRC.cohort, patients = 'adenoma_3')
}
