% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{plot.rev_cohort}
\alias{plot.rev_cohort}
\title{Plot data and trees for a REVOLVER cohort.}
\usage{
\method{plot}{rev_cohort}(x, patients = x$patients, max.phylogenies = 12,
  cex = 1)
}
\arguments{
\item{x}{An object of class \code{"rev_cohort"}.}

\item{patients}{The patients to plot, default is all the one available.}

\item{max.phylogenies}{How many trees should be computed for each patient.}

\item{cex}{Scale cex for graphics.}
}
\value{
nothing
}
\details{
Iterative plotting functions that scans a cohort and runs \code{\link{revolver_report_patient}}
on a set of patients (default all).
}
\examples{
data(Breast.fit)
plot.rev_cohort(Breast.fit, patients = Breast.fit$patients[1:10])
}
