% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_cohort.R
\name{plot.rev_cohort}
\alias{plot.rev_cohort}
\title{Cohort summary scatterplot.}
\usage{
\method{plot}{rev_cohort}(x, cex = 1, ...)
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{cex}{Cex of the plot.}
}
\value{
A \code{ggplot2} object of the plot.
}
\description{
Cohort summary scatterplot.
}
\details{
Returns a scatterplot of the tumour mutational burden. at the
clonal and subclonal level. Each dot is sized by the number of
clones with drivers, and coloured by the numnber of drivers.
}
\examples{
data(Breast.fit)
}
