% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plot_functions.R
\name{plot.rev_cohort}
\alias{plot.rev_cohort}
\title{Summary scatterplot of a cohort.}
\usage{
\method{plot}{rev_cohort}(x, ...)
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{...}{Standard S3 signature.}
}
\value{
A \code{ggplot} object of the plot.
}
\description{
Summary scatterplot of a cohort.
}
\details{
Returns a scatterplot of the tumour mutational burden,  at the
clonal and subclonal level. Each dot is sized by the number of
clones with drivers, and coloured by the numnber of drivers.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

# Cancel the fits - just set the field to NULL
TRACERx_NEJM_2017_REVOLVER$fit = NULL

plot(TRACERx_NEJM_2017_REVOLVER)
}
\seealso{
Other S3 functions: \code{\link{plot.rev_cohort_fit}},
  \code{\link{print.rev_cohort_fit}},
  \code{\link{print.rev_cohort}}
}
\concept{S3 functions}
