% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plot_functions.R
\name{plot.rev_cohort_fit}
\alias{plot.rev_cohort_fit}
\title{Summary scatterplot of a cohort's fits.}
\usage{
\method{plot}{rev_cohort_fit}(x, ...)
}
\arguments{
\item{x}{A REVOLVER cohort with fits.}

\item{...}{Standard S3 signature.}
}
\value{
A \code{ggplot} object of the plot.
}
\description{
Summary scatterplot of a cohort's fits.
}
\details{
Returns a scatterplot of the penalty of a patient's best fit,
versus the tumour mutational burden. Each dot is sized by the number of
combinations of Information Transfer for a patient, and coloured by the
number of annotated drivers.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

plot(TRACERx_NEJM_2017_REVOLVER)
}
\seealso{
Other S3 functions: \code{\link{plot.rev_cohort}},
  \code{\link{print.rev_cohort_fit}},
  \code{\link{print.rev_cohort}}
}
\concept{S3 functions}
