% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_fit.R
\name{plot.rev_cohort_fit}
\alias{plot.rev_cohort_fit}
\title{Plot fits from a REVOLVER cohort.}
\usage{
\method{plot}{rev_cohort_fit}(x, patients = x$patients,
  merge.files = FALSE, cex = 1)
}
\arguments{
\item{x}{An object of class \code{"rev_cohort_fit"}}

\item{patients}{The patients to plot, default is all the one available.}

\item{merge.files}{Merge the output files? (FALSE)}

\item{cex}{Scale cex for graphics}
}
\value{
nothing
}
\description{
Plot fits from a REVOLVER cohort.
}
\details{
Iterative plotting functions that scans a cohort and runs \code{\link{revolver_report_fit_patient}}
on a set of patients (default all).
}
\examples{
data(Breast.fit)
plot.rev_cohort_fit(Breast.fit, patients = Breast.fit$patients[1:4])
}
