% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_cohort_fit.R
\name{plot.rev_cohort_fit}
\alias{plot.rev_cohort_fit}
\title{Fits summary scatterplot.}
\usage{
\method{plot}{rev_cohort_fit}(x, cex = 1, ...)
}
\arguments{
\item{x}{A REVOLVER cohort with fits.}

\item{cex}{Cex of the plot.}
}
\value{
A \code{ggplot2} object of the plot.
}
\description{
Fits summary scatterplot.
}
\details{
Returns a scatterplot of the penalty of a patient's best fit,
versus the tumour mutational burden. Each dot is sized by the number of 
combinations of Information Transfer for a patient, and coloured by the
number of annotated drivers.
}
\examples{
data(Breast.fit)
revolver_plt_penalty_barplot(Breast.fit)
}
