% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_plot_fit.R, R/rev_fit.R
\name{plot.rev_cohort_fit}
\alias{plot.rev_cohort_fit}
\title{Plot REVOLVER fits for a subset of patients in the cohort.}
\usage{
\method{plot}{rev_cohort_fit}(x, patients = x$patients,
  merge.files = FALSE, cex = 1)

\method{plot}{rev_cohort_fit}(x, patients = x$patients,
  merge.files = FALSE, cex = 1)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object}

\item{patients}{The patients to plot, default is all the one available.}

\item{merge.files}{Merge the output files? (FALSE)}

\item{cex}{Scale cex for graphics}

\item{plot.stat}{Set it to \code{TRUE} to visualize also some statistics about each tree}

\item{palette}{RColorBrewer palette for colours of each model.}

\item{out.file}{Output PDF file, default \code{"REVOLVER-cohort.pdf"}.}

\item{layout}{The layout of to arrange the plots.}

\item{alpha}{Transparency}

\item{verbose}{Verbose output.}

\item{...}{Extra parameters}

\item{x}{An object of class \code{"rev_cohort_fit"}}

\item{patients}{The patients to plot, default is all the one available.}
}
\value{
None

nothing
}
\description{
Plot REVOLVER fits for a subset of patients in the cohort.

Plot fits from a REVOLVER cohort.
}
\details{
This function performs the plots for each one of a set of input patient IDs. Each plot
contain the tree fit, its expanded version and the information transfer.

Iterative plotting functions that scans a cohort and runs \code{\link{revolver_report_fit_patient}}
on a set of patients (default all).
}
\examples{
#' data(CRC.cohort)
fit = revolver_fit(CRC.cohort)
plot(fit, "adenoma_3")
data(Breast.fit)
plot.rev_cohort_fit(Breast.fit, patients = Breast.fit$patients[1:4])
}
