% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_phylogeny.R
\name{plot.rev_phylo}
\alias{plot.rev_phylo}
\title{Plot a phylogenetic or a mutation tree for a patient.}
\usage{
\method{plot}{rev_phylo}(x, file = NA, edge.width = NA, edge.label = NA,
  palette = "Set1", graph.cex = 1, table.cex = 1, alpha = 0.7,
  plot.stat = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"rev_phylo"}.}

\item{file}{Output file, or \code{NA}.}

\item{edge.width}{Edge width}

\item{edge.label}{Edge label}

\item{palette}{RColorBrewer palette to colour clusters.}

\item{graph.cex}{Cex for the graph.}

\item{table.cex}{Cex for the table.}

\item{alpha}{Transparency.}

\item{plot.stat}{\code{TRUE} if you want to show some statistics for the tree.}

\item{verbose}{Output.}

\item{...}{Extra parameters}
}
\value{
Nothing
}
\description{
Plot a phylogenetic or a mutation tree for a patient.
}
\examples{
data(CRC.cohort)
plot(CRC.cohort$phylogenies[['adenoma_3']][[1]])
}
