% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_phylogeny.R
\name{plot.rev_phylo}
\alias{plot.rev_phylo}
\title{S3 method that plots a REVOLVER tree.}
\usage{
\method{plot}{rev_phylo}(x, cex = 1,
  node_palette = colorRampPalette(RColorBrewer::brewer.pal(n = 9,
  "Set1")), tree_layout = "tree", information_transfer = FALSE,
  icon = FALSE, ...)
}
\arguments{
\item{x}{A REVOLVER tree (object of class \code{"rev_phylo"}).}

\item{cex}{Cex for the plot.}

\item{node_palette}{A function that applied to a number will return a set of colors.
By default this is a \code{colorRampPalette} applied to 9 colours of the \code{RColorBrewer}
palette \code{Set1}. Colors are generated following a topological sort of the information
transfer, which is obtained from \code{igraph}.}

\item{tree_layout}{A layout that can be used by \code{tidygraph}, which wraps \code{igraph}'s
layouts. By default this is a `tree` layout.}

\item{information_transfer}{If `TRUE`, the plot will show only the information
transfer of the tree. The colouring of the nodes of the trees will match the colouring of
the drivers.}

\item{icon}{If `TRUE` the icon tree version of a tree is plot. This type of view does not show 
a clone unless it has a driver annotated. If this option is true together with \code{information_transfer}
an erorr is generated.}

\item{...}{Extra parameters}
}
\value{
The plot. If `add_information_transfer = TRUE` the object is a combined figure from
package \code{ggpubr}, otherwise it is a single \code{ggplot} object.
}
\description{
S3 method that plots a REVOLVER tree.
}
\examples{
data(CRC.cohort)
plot(CRC.cohort$phylogenies[['adenoma_3']][[1]])
}
