% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CCF_histogram.R
\name{plot_CCF_histogram}
\alias{plot_CCF_histogram}
\title{Plot the data histogram for a patient.}
\usage{
plot_CCF_histogram(x, patient)
}
\arguments{
\item{x}{A REVOLVER cohort object.}

\item{patient}{The id of a patient.}
}
\value{
A \code{ggplot} plot.
}
\description{
If this cohort is working with CCF data, this function plots
the CCF histogram of each one of the samples available for
a patient, facetting the sample variable. 

Otherwise, this plot does not make much sense.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')
 
plot_CCF_histogram(TRACERx_NEJM_2017_REVOLVER, 'CRUK0001')

plot_CCF_histogram(TRACERx_NEJM_2017_REVOLVER, 'CRUK0002')
}
\seealso{
Other Plotting functions: \code{\link{distinct_palette_few}},
  \code{\link{distinct_palette_many}},
  \code{\link{gradient_palette}},
  \code{\link{plot_DET_index}},
  \code{\link{plot_drivers_clonality}},
  \code{\link{plot_drivers_occurrence}},
  \code{\link{plot_graph_drivers}},
  \code{\link{plot_patient_data}},
  \code{\link{plot_patient_mutation_burden}},
  \code{\link{plot_patient_oncoprint}},
  \code{\link{plot_trees_scores}}
}
\concept{Plotting functions}
