% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DET_index.R
\name{plot_DET_index}
\alias{plot_DET_index}
\title{Plot the DET index.}
\usage{
plot_DET_index(x, ...)
}
\arguments{
\item{x}{A \code{REVOLVER} object with fits.}

\item{...}{Parmeters passed to function \code{DET_index} in order to compute
the index value.}
}
\value{
A `ggplot` object of the plot.
}
\description{
Plot the index of Divergent Evolutionary Trajectories, for a set of drivers
using function \code{DET_index}. The plot is a barplot with colours reflecting
the number of distinct incoming edges in each driver (species), and the height
reflecting the actual DET index value.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

plot_DET_index(TRACERx_NEJM_2017_REVOLVER)

# Passing parameters to DET_index
plot_DET_index(TRACERx_NEJM_2017_REVOLVER, min.occurrences = 5)
}
\seealso{
Other Plotting functions: \code{\link{distinct_palette_few}},
  \code{\link{distinct_palette_many}},
  \code{\link{gradient_palette}},
  \code{\link{plot_CCF_histogram}},
  \code{\link{plot_drivers_clonality}},
  \code{\link{plot_drivers_occurrence}},
  \code{\link{plot_graph_drivers}},
  \code{\link{plot_patient_data}},
  \code{\link{plot_patient_mutation_burden}},
  \code{\link{plot_patient_oncoprint}},
  \code{\link{plot_trees_scores}}
}
\concept{Plotting functions}
