% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph_drivers.R
\name{plot_graph_drivers}
\alias{plot_graph_drivers}
\title{Plot summary statistics for the drivers on a graph.}
\usage{
plot_graph_drivers(x, drivers = x$variantIDs.driver,
  min.occurrences = 0, alpha_level = 0.05, cex = 1, ...)
}
\arguments{
\item{x}{A REVOLVER object with fits.}

\item{drivers}{The list of drivers to consider, all by default.
See also function \code{\link{plot_penalty}}.}

\item{min.occurrences}{The minimum number of occurrences for
a trajectory to be considered, zero by default. See also 
function \code{\link{plot_penalty}}.}

\item{alpha_level}{The significance level for the enrichment Fisher test.}

\item{cex}{Cex of the plot.}

\item{...}{Extra parameters passed to the \code{create_layout} function
by \code{ggraph}. For instance, passing \code{algorithm = 'kk'} and 
\code{layout = 'igraph'} the `igraph` layout `kk` will be adopted.}
}
\value{
A `ggplot` object of the plot
}
\description{
Plot a graph with driver genes and annotate with 
different summary statistics for the trajectories that 
involve the drivers. This visualisation shows the frequency
of the driver in the cohort (node size), the penalty for
each pair of odering (edge thickness), the significance for
the pair of orderings as of a Fisher test (edge coloring)
and the overall heterogeneity upstream a driver as of the
DET index (node coloring). This function has parameters
to subset the computation to a list of predefined drivers,
or drivers associated to trajectories with a minimum
recurrence in the fits.
}
