% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trees.R
\name{plot_trees}
\alias{plot_trees}
\title{Plot REVOLVER trees for a patient}
\usage{
plot_trees(x, patient, ...)
}
\arguments{
\item{x}{A REVOLVER cohort object}

\item{patient}{The patient for which the trees should be plot}

\item{...}{Extra parameters, not used.}
}
\value{
A figure assembled with \code{ggpubr}.
}
\description{
This function is like \code{plot_data}, as it uses base plotting
functions to assemble a summary plot for a patient. This function
assembled the plot via \code{ggpubr}. The first line of plots
represents the top tree for a patient, and its information transfer.
The strip below represents up to the top-10 trees for this patient,
as they are obtained from the standard tree-scoring (which means
that the score is not accounting for the actual transfer, but just
for the tree structure).
}
\examples{
TODO
}
