% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trees_scores.R
\name{plot_trees_scores}
\alias{plot_trees_scores}
\title{Plot a barplot of the scored trees for a patient.}
\usage{
plot_trees_scores(x, patient, transfer_palette = distinct_palette_few,
  cex = 1, ...)
}
\arguments{
\item{x}{A REVOLVER cohort object}

\item{transfer_palette}{A function that can be used to sample
an arbitrary number of colours to identify the transfers.}

\item{cex}{Cex of the plot.}

\item{...}{Extra parameters, unused.}

\item{patients}{The plots}
}
\value{
A \code{ggplot} plot.
}
\description{
Returns a barplot where each tree is
a bar, with height proportional to the tree score and
colour to its particular combination of transfer. If
the function is called on a vector of patient ids, the
result is a list of plot.
}
\examples{
data(TRACERx_cohort)

plot_trees_scores(TRACERx_cohort, 'CRUK0002')
}
