% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_cohort.R
\name{print.rev_cohort}
\alias{print.rev_cohort}
\title{Print a \code{"rev_cohort"} object}
\usage{
\method{print}{rev_cohort}(x, ...)
}
\arguments{
\item{x}{obj of class \code{"rev_cohort"}}

\item{digits}{number of output digits}
}
\value{
nothing
}
\description{
Print a \code{"rev_cohort"} object
}
\examples{
data(CRC.cohort)
CRC.cohort
}
