% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_print_functions.R
\name{print.rev_cohort}
\alias{print.rev_cohort}
\title{Summary print of a cohort.}
\usage{
\method{print}{rev_cohort}(x, ...)
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{...}{Standard S3 signature.}
}
\value{
Nothing.
}
\description{
Prints basic information about a cohort object,
status of the information available and other information.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

# Cancel the fits
TRACERx_NEJM_2017_REVOLVER$fit = NULL

print(TRACERx_NEJM_2017_REVOLVER)
}
\seealso{
Other S3 functions: 
\code{\link{plot.rev_cohort_fit}()},
\code{\link{plot.rev_cohort}()},
\code{\link{print.rev_cohort_fit}()}
}
\concept{S3 functions}
