% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_driver.R
\name{remove_driver}
\alias{remove_driver}
\title{Remove a driver event from the cohort.}
\usage{
remove_driver(x, variantID, check = TRUE)
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{variantID}{Id of the driver event to remove.}
}
\value{
A modified cohort where the required event is no longer a driver.
}
\description{
Each event is identied through its id (\code{variantID});
with this function, you can remove a driver events, which consists
in flagging it as \code{FALSE} in the \code{is.Driver} column of the data, and updating
the information transfer. If you have fit the models or clustered the cohort,
you should re-run the analyses after this modification; for this reason,
any previous result from those analyses is cancelled from the returned object.

Notice also that some patients might be removed by this function, because if
they have only one driver then they cannot be fit after driver removal.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

print(TRACERx_NEJM_2017_REVOLVER)

new_cohort = remove_driver(TRACERx_NEJM_2017_REVOLVER, "MET")
print(new_cohort)
}
