% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{revolver_check_cohort}
\alias{revolver_check_cohort}
\title{Check basic inconsistencies in a REVOLVER cohort.}
\usage{
revolver_check_cohort(x, auto.fix = FALSE, return.value = FALSE)
}
\arguments{
\item{x}{An object of class \code{"rev_cohort"}}

\item{auto.fix}{Unused parameter; ideally we will implement some automatic fixing one day.}

\item{return.value}{If the method should return or not a report with the errors. By default is just prints to screen}
}
\value{
depends on \code{return.value}
}
\details{
Perform some basic diagnostic of a cohort object. It will inform of patients without drivers and other
information that can be used to reshape the data before fitting a model.
}
\examples{
data(CRC.cohort)
revolver_check_cohort(CRC.cohort)
print(CRC.cohort) # calls this anyway
}
