% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_cluster.R
\name{revolver_cluster}
\alias{revolver_cluster}
\title{Compute hierarchical clustering for a REVOLVER cohort}
\usage{
revolver_cluster(x, patients = x$patients, hc.method = "ward",
  split.method = "cutreeHybrid", min.group.size = 2)
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object for which the evolutionary distance has been computed.}

\item{hc.method}{Method for hierarchial clustering, see \code{\link{revolver_infoclustering}}.}

\item{split.method}{Method to cut the dendrogram, see \code{\link{revolver_infoclustering}}.}

\item{min.group.size}{Minimum group size for \code{dynamicTreeCut} functions.}
}
\value{
The input \code{x} with a modified field \code{cluster} with results.
}
\description{
Compute hierarchical clustering for a REVOLVER cohort
}
\details{
Compute hierarchical clustering for a REVOLVER cohort with fit models.
You might want to see functions \code{\link{revolver_evo_distance}} and
\code{\link{revolver_infoclustering}} to first compute REVOLVER's
evolutionary distance, and inspect basic parameter values and the clusters
that they allow to identify.
}
\examples{
data(CRC.cohort)
fit = revolver_fit(CRC.cohort)
fit = revolver_evo_distance(fit)
fit = revolver_cluster(fit) # dumped also to disk
}
