% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{revolver_compute_CLtrees}
\alias{revolver_compute_CLtrees}
\title{Compute/ add mutation trees to a REVOLVER cohort.}
\usage{
revolver_compute_CLtrees(x, patient, precomputed.trees = NULL,
  precomputed.scores = NULL, options = list(sspace.cutoff = 10000,
  n.sampling = 5000, overwrite = FALSE, store.max = 100),
  verbose = FALSE)
}
\arguments{
\item{x}{An object of class \code{"rev_cohort"}}

\item{patient}{A patient in the cohort, for which mutation trees are created}

\item{precomputed.trees}{See \code{\link{revolver_compute_phylogenies}}}

\item{precomputed.scores}{See \code{\link{revolver_compute_phylogenies}}}

\item{options}{See \code{\link{revolver_compute_phylogenies}}.}

\item{verbose}{output type.}
}
\value{
a modififed object of class \code{"rev_cohort"} with available
mutation trees for \code{patient}.
}
\description{
Compute/ add mutation trees to a REVOLVER cohort.
}
\details{
This is the analogous of \code{\link{revolver_compute_phylogenies}}, but for binary data and
hence it computes Chow-Liu trees using also Suppes' conditions. Parameters have exactly
the same meaning of the ones described in \code{\link{revolver_compute_phylogenies}}.
}
\examples{
data(CRC.cohort)
new.cohort = revolver_compute_CLtrees(CRC.cohort, patient = 'adenoma_1')
print(new.cohort$phylogenies$adenoma_1)
}
