% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{revolver_compute_phylogenies}
\alias{revolver_compute_phylogenies}
\title{Compute/ add CCF-based phylogenies to a REVOLVER cohort.}
\usage{
revolver_compute_phylogenies(x, patient, precomputed.trees = NULL,
  precomputed.scores = NULL, options = list(sspace.cutoff = 10000,
  n.sampling = 5000, overwrite = FALSE, store.max = 100),
  verbose = FALSE)
}
\arguments{
\item{x}{An object of class \code{"rev_cohort"}}

\item{patient}{A patient in the cohort, for which phylogenies are created}

\item{precomputed.trees}{If a list a of precomputed trees is available, this list should contain their
adjacency matrix. No computation will be carried out in this case.}

\item{precomputed.scores}{If a list a of precomputed trees is available, this list should contain their
scores. No computation will be carried out in this case.}

\item{options}{If one wants to generate new phylogenies for this patient, there named parameters should be
passed through this list. \code{sspace.cutoff = 10000} is the state-space cutoff to generate trees in an
exhaustive fashion, or to Montecarlo sample them. If Montecarlo is chosen \code{sampling} tells how
many trees are sampled. \code{overwrite = FALSE} sets if the patient's current phylogenies
should be overwritten, in case they are already available. \code{store.max = 100} tells how many trees should
be stored, if there are more than \code{store.max} available (ranked).}

\item{verbose}{output type.}
}
\value{
a modififed object of class \code{"rev_cohort"} with available phylogeneis for \code{patient}.
}
\description{
Compute/ add CCF-based phylogenies to a REVOLVER cohort.
}
\details{
Create or compute phylogenies for REVOLVER fit from CCF data. This method should be used also if one
has already pre-computed trees to use for fitting.
}
\examples{
\dontrun{
 TODO
}

}
