% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_dumpStatistics.R
\name{revolver_dumpStatistics}
\alias{revolver_dumpStatistics}
\title{Dump statistics for fit to an Excel file.}
\usage{
revolver_dumpStatistics(x, file = "REVOLVER-Statistics.xlsx")
}
\arguments{
\item{x}{REVOLVER cohort fit}

\item{file}{output file}
}
\value{
invisibly, the list of data.frames written to the Excel file
}
\description{
Dump statistics for fit to an Excel file.
}
\examples{
data(CRC.cohort)
fit <- revolver_fit(CRC.cohort)
revolver_dumpStatistics(fit)

}
