% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_fit.R
\name{revolver_fit}
\alias{revolver_fit}
\title{REVOLVER fit function.}
\usage{
revolver_fit(x, initial.solution = 1, max.iterations = 10,
  restarts = 0, parallel = FALSE, cores.ratio = 0.8,
  transitive.orderings = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{A \code{"rev_cohort"} object with available trees.}

\item{initial.solution}{Either a scalar to fix one initial condition (rank id), or \code{NA} to sample it randomly.}

\item{max.iterations}{Maximum number of EM steps before forcing stop.}

\item{restarts}{Number of initial conditions sampled to compute optimal fit}

\item{parallel}{Use a parallel engine with \code{TRUE}, hides ouptut. Exploits \code{doParallel}.}

\item{cores.ratio}{Percentage of the number of cores to use, after the overall number of
available cores has been detected by \code{doParallel}.}

\item{transitive.orderings}{If transitive orderings should be computed before the second
fitting step (transfering of orderings). The default is \code{FALSE}, which we suggest.}

\item{verbose}{Verbose output}
}
\value{
A \code{"rev_cohort_fit"} object
}
\description{
REVOLVER fit function.
}
\details{
The main fitting function for REVOLVER, implements the 2-steps EM-algorithm described
in the main paper.
}
\examples{
data(CRC.cohort)
fit = revolver_fit(CRC.cohort)
}
