% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_clustering.R
\name{revolver_infoclustering}
\alias{revolver_infoclustering}
\title{Compute possible clusterings for standard parameters.}
\usage{
revolver_infoclustering(x, methods = c("average", "single", "complete",
  "ward"), min.group.size = 2, cex = 1, do.plot = FALSE,
  file = ifelse(!do.plot, NA, "REVOLVER-infoClustering.pdf"))
}
\arguments{
\item{x}{A \code{"rev_cohort_fit"} object for which the evolutionary distance has been computed.}

\item{methods}{Dendogram methods supported by agnes and cluster packages. Default
are all of the following: \code{"average"}, \code{"single"}, \code{"complete"}, \code{"ward"}.}

\item{min.group.size}{Minimum group size for \code{dynamicTreeCut} functions.}

\item{cex}{Cex for the plot}

\item{do.plot}{\code{TRUE} if you want plots to be saved to \code{file}}

\item{file}{PDF output file.}
}
\value{
None, this function just produces a report to be inspected offline.
}
\description{
Compute possible clusterings for standard parameters.
}
\details{
This requires to have computed the distance via function \code{\link{revolver_evo_distance}}.
Compute possible clusterings for standard parameters. This helps
to determine and select clusters in an output dendrogram. This function uses
dendrogram cutting strategies available through packages such as \code{cluster} and
\code{dynamicTreeCut}.
}
\examples{
data(CRC.cohort)
fit = revolver_fit(CRC.cohort)
fit = revolver_evo_distance(fit)
revolver_infoclustering(fit) # dumped to disk
}
