% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revolver_phylogeny.R
\name{revolver_phylogeny}
\alias{revolver_phylogeny}
\title{REVOLVER constructor for a tree}
\usage{
revolver_phylogeny(x, patient, M, score,
  annotation = paste0("A tree for patient ", patient))
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{patient}{The id of this patient in the cohort.}

\item{M}{The tree adjacency matrix.}

\item{score}{A score for this model, that we seek to maximize.}

\item{annotation}{A string to annotate this tree.}
}
\value{
An object of class \code{"rev_phylo"} that represents this tree.
}
\description{
This constructor creates an object of type "rev_phylo", which represents a model
for a patient. This is the same type of object regarldess data are CCF or binary values.
}
\note{
This function is meant to be called after a REVOLVER cohort object has been
created, via the constructure \code{\link{revolver_cohort}}.
}
\examples{

# We re-build the trees for one patient
data(TRACERx_cohort)

# Because the function will return a new cohort, we can just
M = ...

# Create a new tree, the score is made up
new_tree = revolver_phylogeny(TRACERx_cohort, patient = "CRUK0002", M, score = 17)

# S3 functions for this object
print(new_tree)
plot(new_tree)
}
