% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_phylogeny.R
\name{revolver_phylogeny}
\alias{revolver_phylogeny}
\title{REVOLVER constructor for an object of type "rev_phylo": a tree.}
\usage{
revolver_phylogeny(M, patient, dataset, samples, score, annotation = "")
}
\arguments{
\item{M}{Input can be either an adjacency matrix, a dataframe with columns "from" and 'to' that
lists the edges of the model, or a list of edges (vector) in the format "A~B" to denote edge A --> B.}

\item{patient}{Patient ID for this tree.}

\item{dataset}{Data associated to this patient.}

\item{samples}{Thde IDs for the samples of this patient (e.g., "R1", "R2", ...)}

\item{score}{A score for this model, that we seek to maximize.}

\item{annotation}{A string to annotate this tree.}
}
\value{
An object of class "rev_phylo"
}
\details{
REVOLVER constructor to create an object of type "rev_phylo", which represents a model
for a patient. This is the same object regarldess data are CCF or binary values.
}
\examples{
data(CRC)
# This has just one cluster -- we make up its adj_matrix for the tree
dataset = CRC[CRC$patientID == 'adenoma_3', ]
samples_ID = paste('R', 1:6, sep = '')

m = matrix(1, ncol = 1, nrow = 1)
colnames(m) = rownames(m) = '1'
revolver_phylogeny(m, 'adenoma_1', dataset, samples_ID, 52)

}
