% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_phylogeny.R
\name{revolver_phylogeny}
\alias{revolver_phylogeny}
\title{REVOLVER constructor for an object of type "rev_phylo": a tree.}
\usage{
revolver_phylogeny(M, patient, dataset, samples, score, annotation = "")
}
\arguments{
\item{M}{Input can be either an adjacency matrix, a dataframe with columns "from" and 'to' that
lists the edges of the model, or a list of edges (vector) in the format "A~B" to denote edge A --> B.}

\item{patient}{Patient ID for this tree.}

\item{dataset}{Data associated to this patient.}

\item{samples}{Thde IDs for the samples of this patient (e.g., "R1", "R2", ...)}

\item{score}{A score for this model, that we seek to maximize.}

\item{annotation}{A string to annotate this tree.}
}
\value{
An object of class \code{"rev_phylo"}
}
\description{
REVOLVER constructor for an object of type "rev_phylo": a tree.
}
\details{
REVOLVER constructor to create an object of type "rev_phylo", which represents a model
for a patient. This is the same object regarldess data are CCF or binary values.
}
\examples{
# Take data from CRC, inspect sample and dataset
data("CRC")
dataset = CRC[CRC$patientID == 'adenoma_3', ]
samples = paste('R', 1:5, sep = '')

# Extract CCF values
CCF = sapply(dataset$CCF, revolver:::CCF.parser)
CCF = t(apply(CCF, 2, as.numeric))
rownames(CCF) = rownames(dataset)
colnames(CCF) = samples

# Bind a dataset with explicit data
dataset = cbind(dataset, CCF)

# Create empty adj_mat
m = matrix(0, ncol = 4, nrow = 4)
colnames(m) = rownames(m) = 1:4
# Phylogeny: will attach a GL (germline) to the root
# of the tree. In this case all nodes are roots.
revolver_phylogeny(m, 'adenoma_1', dataset, samples, 52)
}
