% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_cohort.R
\name{revolver_removeDriver}
\alias{revolver_removeDriver}
\title{Remove a driver events associated to a certain ID, in a patient.}
\usage{
revolver_removeDriver(x, patient, variantID, misc, cluster)
}
\arguments{
\item{x}{An object of class "rev_cohort"}

\item{patient}{Patient ID.}

\item{variantID}{ID of the event to mark as FALSE.}

\item{misc}{misc of the event to mark as FALSE.}

\item{cluster}{cluster of the event to mark as FALSE.}
}
\value{
see "return.value"
}
\details{
Basic editing function. Each event is identied through its ID,
cluster assignment and "misc" flag. With this function, you can remove it.
The event is not physically removed from the dataset, but instead its flag
is.Driver is set to FALSE. If there are phylogenies inside, they are updated
as well. If you have fit those, however, you should re-run the fit after this
modification.
}
\examples{
data(CRC.cohort)
head(CRC.cohort$dataset) # Info to use to remove
new.data = revolver_removeDriver(CRC.cohort, "adenoma_1", "APC", "NOTHING", "1")
head(new.data$dataset)
}
