% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Samples}
\alias{Samples}
\title{Extract samples name for a patient}
\usage{
Samples(x, p)
}
\arguments{
\item{x}{a REVOLVER object of class "rev_cohort" or "rev_cohort_fit"}

\item{p}{a patient}
}
\value{
CCF/binary data for \code{p}
}
\description{
Extract samples name for a patient
}
\examples{
data(TRACERx_cohort)
Samples(TRACERx_cohort, 'CRUK0001')
}
