% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{Stats}
\alias{Stats}
\title{Summary statistics for the cohort's patients.}
\usage{
Stats(x, patients = x$patients)
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{patients}{The patients for which the summaries are computed.}
}
\value{
A tibble with summary stastics.
}
\description{
Returns the number of samples per patient, the number
of drivers, the number of clonal and subclonal mutations etc. The
function can be run on a subset of patients.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

# Get the stats for all patients
Stats(TRACERx_NEJM_2017_REVOLVER) 

# And subset the patients
Stats(TRACERx_NEJM_2017_REVOLVER, patients = c('CRUK0001', 'CRUK0002')) 
}
\seealso{
Other Summary statistics: \code{\link{Stats_cohort}},
  \code{\link{Stats_drivers}}, \code{\link{Stats_fits}},
  \code{\link{Stats_trees}}
}
\concept{Summary statistics}
