% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Stats_cohort}
\alias{Stats_cohort}
\title{See \code{Stats} function.}
\usage{
Stats_cohort(...)
}
\arguments{
\item{...}{Parameters forwarded to the \code{Stats} function.}
}
\value{
A tibble with summary stastics.
}
\description{
Just a wrapper to the \code{Stats} function.
}
\examples{
data(TRACERx_cohort)

# Get the stats for all patients
Stats_cohort(TRACERx_cohort) 

# And subset the patients
Stats_cohort(TRACERx_cohort, patients = c('CRUK0001', 'CRUK0002')) 
}
