% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{Stats_drivers}
\alias{Stats_drivers}
\title{Summary statistics for the cohort's driver events.}
\usage{
Stats_drivers(x, drivers = x$variantIDs.driver)
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{drivers}{The drivers for which the summaries are computed.}
}
\value{
A tibble with the driver stastics.
}
\description{
Returns the number of clonal and subclonal occurrences
of the drivers in the cohort, and their percentage relative to the
cohort size. The function can be run on a subset of drivers.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

# Get the stats for all patients
Stats_drivers(TRACERx_NEJM_2017_REVOLVER)

# And subset the patients
Stats_drivers(TRACERx_NEJM_2017_REVOLVER, drivers = c('TP53', 'KRAS'))
}
\seealso{
Other Summary statistics: \code{\link{DET_index}},
  \code{\link{Stats_cohort}}, \code{\link{Stats_fits}},
  \code{\link{Stats_trees}}, \code{\link{Stats}}
}
\concept{Summary statistics}
