% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Stats_fits}
\alias{Stats_fits}
\title{Return summary stastics for the cohort's fits}
\usage{
Stats_fits(x, patients = x$patients)
}
\arguments{
\item{x}{A REVOLVER cohort object with fits.}

\item{patients}{The patients for which the summaries are required.}
}
\value{
A tibble with the fits stastics.
}
\description{
Returns a tibble that extends the result of 
\code{Stats_trees} with information about the fit models.
Compared to summaries returns by other \code{Stats_*} functions,
the information from this one is precomputed.
}
\examples{
data(TRACERx_cohort)

# Get the stats for all patients
Stats_fits(TRACERx_cohort) 

# And subset the patients
Stats_fits(TRACERx_cohort, patients = c('CRUK0001', 'CRUK0002')) 
}
