% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_functions.R
\name{Stats_trees}
\alias{Stats_trees}
\title{Summary statistics for the cohort's trees.}
\usage{
Stats_trees(x, patients = x$patients)
}
\arguments{
\item{x}{A REVOLVER cohort.}

\item{patients}{The patients for which the summaries are computed.}
}
\value{
A tibble with the driver stastics.
}
\description{
Returns the statistics about the trees that are available
in the cohort. The function can be run on a subset of patients.
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

# Get the stats for all patients
Stats_trees(TRACERx_NEJM_2017_REVOLVER)

# And subset the patients
Stats_trees(TRACERx_NEJM_2017_REVOLVER, patients = c('CRUK0001', 'CRUK0002'))
}
\seealso{
Other Summary statistics: \code{\link{DET_index}},
  \code{\link{Stats_cohort}}, \code{\link{Stats_drivers}},
  \code{\link{Stats_fits}}, \code{\link{Stats}}
}
\concept{Summary statistics}
