% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter_functions.R
\name{Truncal}
\alias{Truncal}
\title{Extract clonal (i.e., truncal) data for a patient.}
\usage{
Truncal(x, p)
}
\arguments{
\item{x}{A \code{REVOLVER} cohort.}

\item{p}{The id of a patient in the cohort.}
}
\value{
Truncal data for a custom patient.
}
\description{
From a cohort object, this function extracts the tibble of
truncal events for a patient (either CCF or binary).
}
\examples{
# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

Truncal(TRACERx_NEJM_2017_REVOLVER, 'CRUK0002')

Truncal(TRACERx_NEJM_2017_REVOLVER, 'CRUK0008')
}
\seealso{
Other Getters: 
\code{\link{CCF_clusters}()},
\code{\link{CCF}()},
\code{\link{Clonal_cluster}()},
\code{\link{Data}()},
\code{\link{Drivers}()},
\code{\link{ITransfer}()},
\code{\link{Phylo}()},
\code{\link{Samples}()},
\code{\link{Subclonal}()},
\code{\link{get_features}()}
}
\concept{Getters}
